--- next-sitemap.config.js
@@ -1,13 +1,0 @@
-/** @type {import('next-sitemap').IConfig} */
-module.exports = {
-  siteUrl: 'https://earn.superteam.fun/',
-  generateRobotsTxt: true,
-  robotsTxtOptions: {
-    policies: [
-      {
-        userAgent: '*',
-        disallow: ['/signup', '/dashboard', '/new/sponsor', '/t/*/edit'],
-      },
-    ],
-  },
-};
