+++ package.json
@@ -11,7 +11,9 @@
     "check-types": "tsc --noEmit --pretty",
     "prepare": "husky install",
     "postbuild": "next-sitemap",
-    "postinstall": "prisma generate"
+    "postinstall": "prisma generate",
+    "knip": "knip",
+    "find-deadcode": "ts-prune"
   },
   "dependencies": {
     "@bomdi/codebox": "^2.0.0",
@@ -21,7 +23,6 @@
     "@emotion/react": "^11.10.5",
     "@emotion/styled": "^11.10.5",
     "@json2csv/plainjs": "^6.1.3",
-    "@lottiefiles/react-lottie-player": "^3.5.3",
     "@prisma/client": "4.13.0",
     "@project-serum/anchor": "^0.26.0",
     "@solana-mobile/wallet-adapter-mobile": "^2.0.1",
@@ -31,8 +32,6 @@
     "@solana/wallet-adapter-react-ui": "^0.9.33",
     "@solana/wallet-adapter-wallets": "^0.19.20",
     "@solana/web3.js": "^1.75.0",
-    "@tanstack/react-query": "^4.24.4",
-    "@tanstack/react-query-devtools": "^4.26.1",
     "@tiptap/extension-color": "^2.0.0-beta.220",
     "@tiptap/extension-link": "^2.0.0-beta.220",
     "@tiptap/extension-list-item": "^2.0.0-beta.220",
@@ -46,7 +45,6 @@
     "@upstash/qstash": "^2.3.0",
     "@vercel/og": "^0.5.9",
     "airtable": "^0.12.1",
-    "algoliasearch": "^4.16.0",
     "axios": "^1.3.2",
     "base32.js": "^0.1.0",
     "boring-avatars": "^1.7.0",
@@ -57,7 +55,6 @@
     "dayjs": "^1.11.7",
     "degen": "^0.0.121",
     "framer-motion": "^8.5.2",
-    "hamburger-react": "^2.5.0",
     "html-react-parser": "^3.0.12",
     "husky": "^8.0.3",
     "immer": "^9.0.19",
@@ -65,32 +62,25 @@
     "lodash.debounce": "^4.0.8",
     "mixpanel-browser": "^2.47.0",
     "moment": "^2.29.4",
-    "nanoid": "^4.0.2",
     "next": "13.5",
     "next-seo": "^6.0.0",
     "next-sitemap": "^4.0.7",
     "nprogress": "^0.2.0",
-    "open-graph-scraper": "^5.1.1",
     "posthog-js": "^1.93.5",
     "react": "18.2.0",
-    "react-confetti": "^6.1.0",
     "react-countdown": "^2.3.5",
     "react-dom": "18.2.0",
     "react-hook-form": "^7.43.0",
     "react-hot-toast": "^2.4.0",
     "react-icons": "^4.7.1",
     "react-select": "^5.7.0",
-    "react-slick": "^0.29.0",
     "react-textarea-autosize": "^8.5.3",
     "resend": "^0.17.1",
     "sass": "^1.57.1",
     "simple-zustand-devtools": "^1.1.0",
-    "slick-carousel": "^1.8.1",
     "slugify": "^1.6.6",
     "unfurl.js": "^6.3.2",
     "unique-username-generator": "^1.1.3",
-    "use-debounce": "^9.0.4",
-    "uuid": "^9.0.0",
     "zustand": "^4.3.2"
   },
   "devDependencies": {
@@ -99,8 +89,6 @@
     "@types/nprogress": "^0.2.0",
     "@types/react": "18.0.15",
     "@types/react-dom": "18.0.6",
-    "@types/react-slick": "^0.23.10",
-    "@types/uuid": "^9.0.0",
     "@typescript-eslint/eslint-plugin": "^5.48.2",
     "@typescript-eslint/parser": "^5.58.0",
     "eslint": "^8.32.0",
@@ -121,10 +109,12 @@
     "eslint-plugin-simple-import-sort": "^9.0.0",
     "eslint-plugin-testing-library": "^5.9.1",
     "eslint-plugin-unused-imports": "^2.0.0",
+    "knip": "^3.9.0",
     "lint-staged": "^13.2.1",
     "prettier": "^2.8.3",
     "prettier-plugin-prisma": "^4.12.0",
     "prisma": "4.13.0",
+    "ts-prune": "^0.10.3",
     "typescript": "4.9.4"
   }
 }
