--- src/components/listings/listings/winner.tsx
@@ -1,5 +1,0 @@
-import React from 'react';
-
-export const WinnerCard = () => {
-  return <div>WinnerCard</div>;
-};
