--- src/components/misc/confetti.tsx
@@ -1,10 +1,0 @@
-import React from 'react';
-import Confett from 'react-confetti';
-
-export const Confetti = () => {
-  return (
-    <>
-      <Confett gravity={1} numberOfPieces={600} />
-    </>
-  );
-};
