+++ src/constants/index.ts
@@ -1,9 +1,49 @@
-import type { Experience } from '../interface/listings';
+import type { SkillMap } from '@/interface/skills';
 
 export interface MultiSelectOptions {
   value: string;
   label: string;
 }
+
+export const skillMap: SkillMap[] = [
+  {
+    mainskill: 'Growth',
+    color: '#16A368',
+  },
+  {
+    mainskill: 'Frontend',
+    color: '#3E8BFF',
+  },
+  {
+    mainskill: 'Backend',
+    color: '#FF833E',
+  },
+  {
+    mainskill: 'Blockchain',
+    color: '#FF3EC9',
+  },
+  {
+    mainskill: 'Design',
+    color: '#7E51FF',
+  },
+  {
+    mainskill: 'Content',
+    color: '#5EA8BF',
+  },
+  {
+    mainskill: 'Community',
+    color: '#F5C723',
+  },
+  {
+    mainskill: 'Mobile',
+    color: '#7E51FF',
+  },
+  {
+    mainskill: 'Other',
+    color: '#64758B',
+  },
+];
+
 export const IndustryList: MultiSelectOptions[] = [
   {
     label: 'DAOs',
@@ -398,203 +438,8 @@ export const skillSubSkillMap = {
     },
   ],
 };
-export const SubSkills: MultiSelectOptions[] = [
-  {
-    label: 'React',
-    value: 'React',
-  },
-  {
-    label: 'Angular',
-    value: 'Angular',
-  },
-  {
-    label: 'Vue',
-    value: 'Vue',
-  },
-  {
-    label: 'Redux',
-    value: 'Redux',
-  },
-  {
-    label: 'Rust',
-    value: 'Rust',
-  },
-  {
-    label: 'Solidity',
-    value: 'Solidity',
-  },
-  {
-    label: 'Move',
-    value: 'Move',
-  },
-  {
-    label: 'Javascript',
-    value: 'Javascript',
-  },
-  {
-    label: 'PHP',
-    value: 'PHP',
-  },
-  {
-    label: 'Python',
-    value: 'Python',
-  },
-  {
-    label: 'Java',
-    value: 'Java',
-  },
-  {
-    label: 'C++',
-    value: 'C++',
-  },
-  {
-    label: 'C',
-    value: 'C',
-  },
-  {
-    label: 'Ruby',
-    value: 'Ruby',
-  },
-  {
-    label: 'Go',
-    value: 'Go',
-  },
-  {
-    label: 'MySQL',
-    value: 'MySQL',
-  },
-  {
-    label: 'Postgres',
-    value: 'Postgres',
-  },
-  {
-    label: 'MongoDB',
-    value: 'MongoDB',
-  },
-  {
-    label: 'Data Analytics',
-    value: 'Data Analytics',
-  },
-  {
-    label: 'Operations',
-    value: 'Operations',
-  },
-  {
-    label: 'Community Manager',
-    value: 'Community Manager',
-  },
-  {
-    label: 'Discord Moderator',
-    value: 'Discord Moderator',
-  },
-  {
-    label: 'Research',
-    value: 'Research',
-  },
-  {
-    label: 'Video',
-    value: 'Video',
-  },
-  {
-    label: 'Writing',
-    value: 'Writing',
-  },
-  {
-    label: 'Social Media',
-    value: 'Social Media',
-  },
-  {
-    label: 'Business Development',
-    value: 'Business Development',
-  },
-  {
-    label: 'Digital Marketing',
-    value: 'Digital Marketing',
-  },
-  {
-    label: 'Marketing',
-    value: 'Marketing',
-  },
-  {
-    label: 'UI/UX Design',
-    value: 'UI/UX Design',
-  },
-  {
-    label: 'Graphic Design',
-    value: 'Graphic Design',
-  },
-  {
-    label: 'Illustration',
-    value: 'Illustration',
-  },
-  {
-    label: 'Game Design',
-    value: 'Game Design',
-  },
-  {
-    label: 'Presentation Design',
-    value: 'Presentation Design',
-  },
-  {
-    label: 'Android',
-    value: 'Android',
-  },
-  {
-    label: 'iOS',
-    value: 'iOS',
-  },
-  {
-    label: 'Flutter',
-    value: 'Flutter',
-  },
-  {
-    label: 'React Native',
-    value: 'React Native',
-  },
-  {
-    label: 'Other',
-    value: 'Other',
-  },
-];
-export const ExperienceList: Experience[] = [
-  '0 Yrs: Fresher/Graduate ',
-  '0-1 Yrs: Some Experience Required',
-  '1-5 Yrs: Early Career Professional',
-  '5-10 Yrs: Mid Career Professional',
-  '10 Yrs+: Senior Professional',
-];
-export const TimeZoneList: MultiSelectOptions[] = [
-  {
-    label: 'Asia',
-    value: 'Asia',
-  },
-  {
-    label: 'India',
-    value: 'India',
-  },
-  {
-    label: 'North/South America',
-    value: 'North/South America',
-  },
-  {
-    label: 'APAC',
-    value: 'APAC',
-  },
-  {
-    label: 'EMEA',
-    value: 'EMEA',
-  },
-  {
-    label: 'All',
-    value: 'All',
-  },
-];
-export const PrizeList = ['first', 'second', 'third', 'fourth', 'fifth'];
 
-export const ListingTypeQueryMap = {
-  Bounties: 'bounties',
-  Grants: 'grants',
-};
+export const PrizeList = ['first', 'second', 'third', 'fourth', 'fifth'];
 
 export const CountryList: string[] = [
   'India',
@@ -855,22 +700,6 @@ export const CommunityList: string[] = [
   '10K Designers',
   'Other',
 ];
-
-export const CommunityImage: { [key in string]: string } = {
-  Superteam: '/assets/talent/superteam-logo.png',
-  LamportDAO:
-    'https://white-chemical-flamingo-683.mypinata.cloud/ipfs/QmSdqwxpmr2ouFFTapMbywqiBoEzJ3NBUEsAAZ8szB72Sj',
-  "Grape / Dean's List":
-    'https://white-chemical-flamingo-683.mypinata.cloud/ipfs/QmWQEuQiVMBcdw8DgdddvbZ4d2s4xyoy1aCCmrkeRj8tGk',
-  SuperWomenDao:
-    'https://white-chemical-flamingo-683.mypinata.cloud/ipfs/QmPiU1v6mLdn37kikdJ3NRfjJdTFRf26Tv6cAU7veHx3EA',
-  DeveloperDAO:
-    'https://white-chemical-flamingo-683.mypinata.cloud/ipfs/QmQUKa7gHjsdpqk7Vt2z5ZSA4P6Kgn3CRBkTvWquNVo7w3',
-  Metacamp:
-    'https://white-chemical-flamingo-683.mypinata.cloud/ipfs/QmUzZWzWhAZuuMPjnwcqLbbyJJB48NhaaA78cFTyzbkGZc',
-  '10K Designers':
-    'https://white-chemical-flamingo-683.mypinata.cloud/ipfs/QmafVJ8P5yNxzZmLxX3EY28vNwy1snYv7JHErCuTgmR5cs',
-};
 
 export const web3Exp = [
   'New to crypto',
