--- src/interface/seo.ts
@@ -1,10 +1,0 @@
-export type SeoPropsType = {
-  title: string;
-  description: string;
-  image: string;
-  name: string;
-  twitterHandle: string;
-  domain: string;
-};
-
-export default SeoPropsType;
