+++ src/utils/airtable.ts
@@ -1,6 +1,6 @@
 import Airtable from 'airtable';
 
-export function getDatabase() {
+function getDatabase() {
   const base = new Airtable({ apiKey: process.env.AIRTABLE_KEY }).base(
     process.env.AIRTABLE_BASE!
   );
