--- src/utils/nft.ts
@@ -1,193 +1,0 @@
-export const genNFT = () => {
-  return `<svg
-  width="565"
-  height="783"
-  viewBox="0 0 565 783"
-  fill="none"
-  xmlns="http://www.w3.org/2000/svg"
->
-  <g clip-path="url(#clip0_41_9577)">
-    <rect width="565" height="783" rx="20" fill="#561CFD" />
-    <path d="M154 270L392 508" stroke="#896BDF" />
-    <path d="M392 270L154 508" stroke="#896BDF" />
-    <path
-      d="M341.652 335.853H386V352.965C386 376.172 367.227 394.929 344.001 394.929H341.652V335.853ZM278.538 311H341.652V472.531H334.142C288.391 472.531 279.24 440.649 279.24 412.278L278.538 311ZM176 347.576C176 376.41 196.646 386.962 219.64 390.479H176V473H217.999C260.235 473 265.159 454.243 265.159 436.424C265.159 414.619 250.145 399.385 226.912 393.521H265.159V311H223.16C180.929 311 176 329.757 176 347.576Z"
-      fill="white"
-    />
-    <path
-      d="M56.0881 124H51.8949L61.3168 97.8182H65.8807L75.3026 124H71.1094L63.7074 102.574H63.5028L56.0881 124ZM56.7912 113.747H70.3935V117.071H56.7912V113.747ZM82.5991 112.341V124H78.7766V104.364H82.4457V107.56H82.6886C83.1403 106.52 83.8477 105.685 84.8107 105.054C85.7823 104.423 87.0053 104.108 88.4798 104.108C89.8178 104.108 90.9897 104.389 91.9954 104.952C93.0011 105.506 93.7809 106.332 94.3349 107.432C94.8889 108.531 95.1658 109.891 95.1658 111.51V124H91.3434V111.97C91.3434 110.547 90.9727 109.435 90.2312 108.634C89.4897 107.824 88.4712 107.419 87.1758 107.419C86.2894 107.419 85.5011 107.611 84.8107 107.994C84.1289 108.378 83.5877 108.94 83.1871 109.682C82.7951 110.415 82.5991 111.301 82.5991 112.341ZM106.007 124.435C104.762 124.435 103.637 124.205 102.632 123.744C101.626 123.276 100.829 122.598 100.241 121.712C99.6616 120.825 99.3718 119.739 99.3718 118.452C99.3718 117.344 99.5849 116.432 100.011 115.716C100.437 115 101.012 114.433 101.737 114.016C102.461 113.598 103.271 113.283 104.166 113.07C105.061 112.857 105.973 112.695 106.902 112.584C108.078 112.447 109.032 112.337 109.765 112.251C110.498 112.158 111.031 112.009 111.363 111.804C111.696 111.599 111.862 111.267 111.862 110.807V110.717C111.862 109.601 111.547 108.736 110.916 108.122C110.294 107.509 109.365 107.202 108.129 107.202C106.842 107.202 105.828 107.487 105.086 108.058C104.353 108.621 103.846 109.247 103.565 109.938L99.9727 109.119C100.399 107.926 101.021 106.963 101.839 106.23C102.666 105.489 103.616 104.952 104.69 104.619C105.764 104.278 106.893 104.108 108.078 104.108C108.862 104.108 109.693 104.202 110.571 104.389C111.457 104.568 112.284 104.901 113.051 105.386C113.826 105.872 114.461 106.567 114.956 107.47C115.45 108.365 115.697 109.528 115.697 110.96V124H111.964V121.315H111.811C111.564 121.81 111.193 122.295 110.699 122.773C110.204 123.25 109.569 123.646 108.794 123.962C108.018 124.277 107.089 124.435 106.007 124.435ZM106.838 121.366C107.895 121.366 108.798 121.158 109.548 120.74C110.306 120.322 110.882 119.777 111.274 119.104C111.674 118.422 111.875 117.693 111.875 116.918V114.386C111.738 114.523 111.474 114.651 111.082 114.77C110.699 114.881 110.26 114.979 109.765 115.064C109.271 115.141 108.789 115.213 108.321 115.281C107.852 115.341 107.46 115.392 107.145 115.435C106.403 115.528 105.725 115.686 105.112 115.908C104.507 116.129 104.021 116.449 103.654 116.866C103.297 117.276 103.118 117.821 103.118 118.503C103.118 119.449 103.467 120.165 104.166 120.651C104.865 121.128 105.755 121.366 106.838 121.366ZM129.955 104.364V107.432H119.229V104.364H129.955ZM122.105 99.6591H125.928V118.234C125.928 118.976 126.038 119.534 126.26 119.909C126.482 120.276 126.767 120.527 127.116 120.663C127.474 120.791 127.862 120.855 128.28 120.855C128.587 120.855 128.855 120.834 129.085 120.791C129.315 120.749 129.494 120.714 129.622 120.689L130.312 123.847C130.091 123.932 129.776 124.017 129.366 124.102C128.957 124.196 128.446 124.247 127.832 124.256C126.827 124.273 125.889 124.094 125.02 123.719C124.151 123.344 123.447 122.764 122.911 121.98C122.374 121.196 122.105 120.212 122.105 119.027V99.6591ZM142.243 124.396C140.402 124.396 138.796 123.974 137.424 123.131C136.051 122.287 134.986 121.107 134.228 119.589C133.469 118.072 133.09 116.3 133.09 114.271C133.09 112.234 133.469 110.453 134.228 108.928C134.986 107.402 136.051 106.217 137.424 105.374C138.796 104.53 140.402 104.108 142.243 104.108C144.084 104.108 145.691 104.53 147.063 105.374C148.435 106.217 149.5 107.402 150.259 108.928C151.017 110.453 151.397 112.234 151.397 114.271C151.397 116.3 151.017 118.072 150.259 119.589C149.5 121.107 148.435 122.287 147.063 123.131C145.691 123.974 144.084 124.396 142.243 124.396ZM142.256 121.188C143.449 121.188 144.438 120.872 145.222 120.241C146.006 119.611 146.586 118.771 146.961 117.723C147.344 116.675 147.536 115.52 147.536 114.259C147.536 113.006 147.344 111.855 146.961 110.807C146.586 109.75 146.006 108.902 145.222 108.263C144.438 107.624 143.449 107.304 142.256 107.304C141.054 107.304 140.057 107.624 139.265 108.263C138.48 108.902 137.897 109.75 137.513 110.807C137.138 111.855 136.951 113.006 136.951 114.259C136.951 115.52 137.138 116.675 137.513 117.723C137.897 118.771 138.48 119.611 139.265 120.241C140.057 120.872 141.054 121.188 142.256 121.188ZM159.486 97.8182V124H155.663V97.8182H159.486ZM167.044 131.364C166.473 131.364 165.953 131.317 165.485 131.223C165.016 131.138 164.667 131.044 164.436 130.942L165.357 127.81C166.056 127.997 166.678 128.078 167.223 128.053C167.769 128.027 168.25 127.822 168.668 127.439C169.094 127.055 169.469 126.429 169.793 125.56L170.266 124.256L163.081 104.364H167.172L172.145 119.602H172.35L177.323 104.364H181.426L173.334 126.621C172.959 127.643 172.482 128.509 171.902 129.216C171.323 129.932 170.632 130.469 169.831 130.827C169.03 131.185 168.101 131.364 167.044 131.364ZM212.572 106.089C212.325 105.314 211.992 104.619 211.575 104.006C211.165 103.384 210.675 102.855 210.104 102.42C209.533 101.977 208.881 101.641 208.148 101.411C207.424 101.18 206.627 101.065 205.758 101.065C204.283 101.065 202.954 101.445 201.769 102.203C200.585 102.962 199.647 104.074 198.957 105.54C198.275 106.997 197.934 108.783 197.934 110.896C197.934 113.018 198.279 114.812 198.969 116.278C199.66 117.744 200.606 118.857 201.808 119.615C203.009 120.374 204.377 120.753 205.911 120.753C207.335 120.753 208.575 120.463 209.631 119.884C210.697 119.304 211.519 118.486 212.099 117.429C212.687 116.364 212.981 115.111 212.981 113.67L214.004 113.862H206.512V110.602H216.803V113.581C216.803 115.78 216.335 117.689 215.397 119.308C214.468 120.919 213.181 122.163 211.536 123.041C209.9 123.919 208.025 124.358 205.911 124.358C203.542 124.358 201.462 123.812 199.673 122.722C197.891 121.631 196.502 120.084 195.505 118.081C194.508 116.07 194.009 113.683 194.009 110.922C194.009 108.834 194.299 106.959 194.879 105.297C195.458 103.635 196.272 102.224 197.32 101.065C198.377 99.8977 199.617 99.0071 201.04 98.3935C202.472 97.7713 204.036 97.4602 205.732 97.4602C207.147 97.4602 208.464 97.669 209.683 98.0866C210.91 98.5043 212.001 99.0966 212.955 99.8636C213.918 100.631 214.715 101.543 215.346 102.599C215.977 103.648 216.403 104.811 216.624 106.089H212.572ZM229.782 124.396C227.941 124.396 226.335 123.974 224.963 123.131C223.591 122.287 222.525 121.107 221.767 119.589C221.008 118.072 220.629 116.3 220.629 114.271C220.629 112.234 221.008 110.453 221.767 108.928C222.525 107.402 223.591 106.217 224.963 105.374C226.335 104.53 227.941 104.108 229.782 104.108C231.623 104.108 233.23 104.53 234.602 105.374C235.974 106.217 237.039 107.402 237.798 108.928C238.556 110.453 238.936 112.234 238.936 114.271C238.936 116.3 238.556 118.072 237.798 119.589C237.039 121.107 235.974 122.287 234.602 123.131C233.23 123.974 231.623 124.396 229.782 124.396ZM229.795 121.188C230.988 121.188 231.977 120.872 232.761 120.241C233.545 119.611 234.125 118.771 234.5 117.723C234.883 116.675 235.075 115.52 235.075 114.259C235.075 113.006 234.883 111.855 234.5 110.807C234.125 109.75 233.545 108.902 232.761 108.263C231.977 107.624 230.988 107.304 229.795 107.304C228.593 107.304 227.596 107.624 226.804 108.263C226.02 108.902 225.436 109.75 225.052 110.807C224.677 111.855 224.49 113.006 224.49 114.259C224.49 115.52 224.677 116.675 225.052 117.723C225.436 118.771 226.02 119.611 226.804 120.241C227.596 120.872 228.593 121.188 229.795 121.188ZM246.718 117.339L246.692 112.673H247.357L255.181 104.364H259.758L250.835 113.824H250.234L246.718 117.339ZM243.202 124V97.8182H247.025V124H243.202ZM255.603 124L248.572 114.668L251.205 111.996L260.295 124H255.603ZM268.956 124.435C267.712 124.435 266.587 124.205 265.581 123.744C264.575 123.276 263.778 122.598 263.19 121.712C262.611 120.825 262.321 119.739 262.321 118.452C262.321 117.344 262.534 116.432 262.96 115.716C263.386 115 263.962 114.433 264.686 114.016C265.411 113.598 266.22 113.283 267.115 113.07C268.01 112.857 268.922 112.695 269.851 112.584C271.027 112.447 271.982 112.337 272.714 112.251C273.447 112.158 273.98 112.009 274.312 111.804C274.645 111.599 274.811 111.267 274.811 110.807V110.717C274.811 109.601 274.496 108.736 273.865 108.122C273.243 107.509 272.314 107.202 271.078 107.202C269.791 107.202 268.777 107.487 268.036 108.058C267.303 108.621 266.795 109.247 266.514 109.938L262.922 109.119C263.348 107.926 263.97 106.963 264.788 106.23C265.615 105.489 266.565 104.952 267.639 104.619C268.713 104.278 269.842 104.108 271.027 104.108C271.811 104.108 272.642 104.202 273.52 104.389C274.406 104.568 275.233 104.901 276 105.386C276.776 105.872 277.411 106.567 277.905 107.47C278.399 108.365 278.646 109.528 278.646 110.96V124H274.913V121.315H274.76C274.513 121.81 274.142 122.295 273.648 122.773C273.153 123.25 272.518 123.646 271.743 123.962C270.967 124.277 270.038 124.435 268.956 124.435ZM269.787 121.366C270.844 121.366 271.747 121.158 272.497 120.74C273.256 120.322 273.831 119.777 274.223 119.104C274.624 118.422 274.824 117.693 274.824 116.918V114.386C274.688 114.523 274.423 114.651 274.031 114.77C273.648 114.881 273.209 114.979 272.714 115.064C272.22 115.141 271.739 115.213 271.27 115.281C270.801 115.341 270.409 115.392 270.094 115.435C269.352 115.528 268.675 115.686 268.061 115.908C267.456 116.129 266.97 116.449 266.604 116.866C266.246 117.276 266.067 117.821 266.067 118.503C266.067 119.449 266.416 120.165 267.115 120.651C267.814 121.128 268.705 121.366 269.787 121.366ZM287.56 97.8182V124H283.738V97.8182H287.56Z"
-      fill="white"
-    />
-    <path
-      d="M52.8377 83V66.2727H63.2106V68.4453H55.3615V73.5419H62.4673V75.7063H55.3615V83H52.8377ZM73.2792 77.7972V70.4545H75.7295V83H73.3282V80.8274H73.1975C72.9089 81.4972 72.4461 82.0553 71.809 82.5018C71.1774 82.9428 70.3906 83.1634 69.4486 83.1634C68.6427 83.1634 67.9294 82.9864 67.3087 82.6325C66.6934 82.2731 66.2088 81.7422 65.8548 81.0398C65.5064 80.3374 65.3321 79.4689 65.3321 78.4343V70.4545H67.7742V78.1403C67.7742 78.9951 68.0111 79.6758 68.4848 80.1822C68.9585 80.6886 69.5738 80.9418 70.3307 80.9418C70.7881 80.9418 71.2427 80.8274 71.6947 80.5987C72.1521 80.37 72.5305 80.0243 72.83 79.5614C73.1349 79.0986 73.2847 78.5105 73.2792 77.7972ZM81.4529 66.2727V83H79.0108V66.2727H81.4529ZM87.1805 66.2727V83H84.7384V66.2727H87.1805ZM110.25 66.2727V83H107.93L99.4278 70.7322H99.2726V83H96.7488V66.2727H99.0848L107.595 78.5568H107.751V66.2727H110.25ZM117.386 83.2777C116.591 83.2777 115.873 83.1307 115.23 82.8366C114.588 82.5372 114.078 82.1043 113.703 81.538C113.333 80.9717 113.147 80.2775 113.147 79.4553C113.147 78.7474 113.283 78.1648 113.556 77.7074C113.828 77.25 114.196 76.8879 114.658 76.6211C115.121 76.3543 115.638 76.1528 116.21 76.0167C116.782 75.8806 117.365 75.7771 117.958 75.7063C118.71 75.6192 119.319 75.5484 119.788 75.494C120.256 75.4341 120.596 75.3388 120.809 75.2081C121.021 75.0774 121.127 74.8651 121.127 74.571V74.5138C121.127 73.8005 120.926 73.2479 120.523 72.8558C120.125 72.4638 119.532 72.2678 118.742 72.2678C117.92 72.2678 117.272 72.4502 116.798 72.815C116.33 73.1744 116.006 73.5746 115.826 74.0156L113.531 73.4929C113.804 72.7306 114.201 72.1153 114.724 71.647C115.252 71.1733 115.859 70.8303 116.545 70.6179C117.231 70.4001 117.953 70.2912 118.71 70.2912C119.21 70.2912 119.741 70.3511 120.302 70.4709C120.868 70.5852 121.397 70.7976 121.887 71.108C122.382 71.4183 122.788 71.8621 123.104 72.4393C123.42 73.011 123.577 73.7543 123.577 74.669V83H121.192V81.2848H121.094C120.937 81.6006 120.7 81.911 120.384 82.2159C120.068 82.5208 119.662 82.774 119.167 82.9755C118.671 83.177 118.078 83.2777 117.386 83.2777ZM117.917 81.3175C118.592 81.3175 119.17 81.1841 119.649 80.9173C120.133 80.6504 120.501 80.302 120.751 79.8718C121.007 79.4362 121.135 78.9706 121.135 78.4751V76.858C121.048 76.9451 120.879 77.0268 120.629 77.103C120.384 77.1738 120.103 77.2364 119.788 77.2908C119.472 77.3398 119.164 77.3861 118.865 77.4297C118.565 77.4678 118.315 77.5005 118.113 77.5277C117.64 77.5876 117.207 77.6883 116.815 77.8299C116.428 77.9715 116.118 78.1757 115.884 78.4425C115.655 78.7038 115.54 79.0523 115.54 79.4879C115.54 80.0923 115.764 80.5497 116.21 80.8601C116.657 81.165 117.226 81.3175 117.917 81.3175ZM126.83 83V70.4545H129.174V72.4964H129.329C129.591 71.8049 130.018 71.2659 130.612 70.8793C131.205 70.4872 131.916 70.2912 132.744 70.2912C133.582 70.2912 134.284 70.4872 134.851 70.8793C135.422 71.2713 135.844 71.8104 136.117 72.4964H136.247C136.547 71.8267 137.023 71.2931 137.677 70.8956C138.33 70.4927 139.109 70.2912 140.013 70.2912C141.151 70.2912 142.079 70.6478 142.798 71.3612C143.522 72.0745 143.884 73.1499 143.884 74.5874V83H141.442V74.8161C141.442 73.9666 141.211 73.3513 140.748 72.9702C140.285 72.589 139.732 72.3984 139.09 72.3984C138.295 72.3984 137.677 72.6435 137.236 73.1335C136.795 73.6181 136.574 74.2416 136.574 75.0039V83H134.14V74.6609C134.14 73.9802 133.928 73.433 133.503 73.0192C133.078 72.6054 132.526 72.3984 131.845 72.3984C131.382 72.3984 130.955 72.521 130.563 72.766C130.176 73.0056 129.863 73.3404 129.623 73.7706C129.389 74.2008 129.272 74.699 129.272 75.2653V83H126.83ZM152.564 83.2532C151.328 83.2532 150.264 82.9891 149.371 82.4609C148.483 81.9273 147.797 81.1786 147.312 80.2148C146.833 79.2456 146.594 78.1103 146.594 76.8089C146.594 75.5239 146.833 74.3913 147.312 73.4112C147.797 72.4311 148.472 71.6661 149.338 71.1161C150.209 70.5662 151.228 70.2912 152.393 70.2912C153.101 70.2912 153.787 70.4083 154.451 70.6424C155.115 70.8765 155.712 71.2441 156.24 71.745C156.768 72.246 157.184 72.8967 157.489 73.6971C157.794 74.4921 157.947 75.4586 157.947 76.5966V77.4624H147.974V75.6328H155.554C155.554 74.9903 155.423 74.4213 155.162 73.9258C154.9 73.4248 154.533 73.0301 154.059 72.7415C153.591 72.4529 153.041 72.3086 152.409 72.3086C151.723 72.3086 151.124 72.4774 150.612 72.815C150.106 73.1471 149.714 73.5827 149.436 74.1218C149.164 74.6554 149.028 75.2353 149.028 75.8615V77.2908C149.028 78.1294 149.175 78.8427 149.469 79.4308C149.768 80.0188 150.185 80.468 150.718 80.7784C151.252 81.0833 151.875 81.2358 152.589 81.2358C153.052 81.2358 153.474 81.1705 153.855 81.0398C154.236 80.9036 154.565 80.7022 154.843 80.4354C155.121 80.1686 155.333 79.8391 155.48 79.4471L157.792 79.8636C157.606 80.5443 157.274 81.1405 156.795 81.6523C156.321 82.1587 155.725 82.5535 155.006 82.8366C154.293 83.1143 153.479 83.2532 152.564 83.2532Z"
-      fill="#BCACE7"
-    />
-    <path
-      d="M122.146 822.223C117.695 822.223 113.425 820.455 110.278 817.307C107.13 814.159 105.362 809.89 105.362 805.439C105.362 800.987 107.13 796.718 110.278 793.571C113.425 790.423 117.695 788.655 122.146 788.655V766.899C111.925 766.899 102.122 770.96 94.8944 778.187C87.6668 785.415 83.6064 795.217 83.6064 805.439C83.6064 815.66 87.6668 825.463 94.8944 832.69C102.122 839.918 111.925 843.978 122.146 843.978V822.223Z"
-      fill="#E067F3"
-    />
-    <path
-      d="M45.0886 822.223C49.5401 822.223 53.8093 820.455 56.9569 817.307C60.1046 814.159 61.8729 809.89 61.8729 805.439C61.8729 800.987 60.1046 796.718 56.9569 793.571C53.8093 790.423 49.5401 788.655 45.0886 788.655V766.899C55.31 766.899 65.1127 770.96 72.3402 778.187C79.5678 785.415 83.6283 795.217 83.6283 805.439C83.6283 815.66 79.5678 825.463 72.3402 832.69C65.1127 839.918 55.31 843.978 45.0886 843.978V822.223Z"
-      fill="#E067F3"
-    />
-    <path
-      d="M137.549 669.012V843.988C160.497 843.602 182.376 834.215 198.468 817.85C214.561 801.485 223.579 779.452 223.579 756.5C223.579 733.548 214.561 711.515 198.468 695.15C182.376 678.785 160.497 669.398 137.549 669.012Z"
-      fill="#E067F3"
-    />
-    <path
-      d="M312.524 718.392V669.012C300.911 668.817 289.376 670.936 278.591 675.245C267.805 679.554 257.986 685.967 249.704 694.111C241.423 702.254 234.846 711.964 230.356 722.676C225.867 733.387 223.554 744.886 223.554 756.5C223.554 768.114 225.867 779.613 230.356 790.324C234.846 801.036 241.423 810.746 249.704 818.889C257.986 827.033 267.805 833.446 278.591 837.755C289.376 842.064 300.911 844.183 312.524 843.988V794.608C307.438 794.738 302.378 793.848 297.641 791.991C292.904 790.134 288.587 787.348 284.944 783.796C281.301 780.245 278.405 776 276.429 771.312C274.452 766.624 273.433 761.588 273.433 756.5C273.433 751.412 274.452 746.376 276.429 741.688C278.405 737 281.301 732.755 284.944 729.204C288.587 725.652 292.904 722.866 297.641 721.009C302.378 719.152 307.438 718.262 312.524 718.392Z"
-      fill="#E067F3"
-    />
-    <path d="M45.0886 669V746.079L122.179 669H45.0886Z" fill="#E067F3" />
-    <path d="M83.6389 746.079L122.179 707.54H83.6389V746.079Z" fill="#E067F3" />
-    <path
-      d="M327.914 669.006V710.876C330.705 710.942 333.482 710.45 336.08 709.428C338.678 708.405 341.045 706.874 343.042 704.923C345.04 702.973 346.627 700.643 347.711 698.07C348.794 695.497 349.353 692.733 349.353 689.941C349.353 687.149 348.794 684.385 347.711 681.813C346.627 679.24 345.04 676.909 343.042 674.959C341.045 673.008 338.678 671.477 336.08 670.454C333.482 669.432 330.705 668.94 327.914 669.006Z"
-      fill="#E067F3"
-    />
-    <path
-      d="M369.784 669.006V710.876C364.318 710.746 359.119 708.484 355.299 704.572C351.479 700.66 349.341 695.409 349.341 689.941C349.341 684.473 351.479 679.223 355.299 675.311C359.119 671.399 364.318 669.136 369.784 669.006Z"
-      fill="#E067F3"
-    />
-    <path
-      d="M411.666 669.006H369.784C369.784 674.56 371.99 679.886 375.918 683.814C379.845 687.741 385.171 689.947 390.725 689.947C396.279 689.947 401.606 687.741 405.533 683.814C409.46 679.886 411.666 674.56 411.666 669.006Z"
-      fill="#E067F3"
-    />
-    <path
-      d="M411.666 710.876H369.784C369.784 705.322 371.99 699.996 375.918 696.069C379.845 692.141 385.171 689.935 390.725 689.935C396.279 689.935 401.606 692.141 405.533 696.069C409.46 699.996 411.666 705.322 411.666 710.876Z"
-      fill="#E067F3"
-    />
-    <path
-      d="M411.666 710.876V752.817C406.112 752.817 400.786 750.611 396.859 746.684C392.931 742.757 390.725 737.43 390.725 731.876C390.725 726.322 392.931 720.996 396.859 717.069C400.786 713.142 406.112 710.935 411.666 710.935V710.876Z"
-      fill="#E067F3"
-    />
-    <path
-      d="M369.784 710.876V752.817C375.338 752.817 380.664 750.611 384.592 746.684C388.519 742.757 390.725 737.43 390.725 731.876C390.725 726.322 388.519 720.996 384.592 717.069C380.664 713.142 375.338 710.935 369.784 710.935V710.876Z"
-      fill="#E067F3"
-    />
-    <path
-      d="M327.914 710.876H369.784C369.851 713.667 369.358 716.443 368.336 719.041C367.314 721.639 365.782 724.007 363.831 726.004C361.881 728.002 359.551 729.589 356.978 730.673C354.405 731.756 351.641 732.315 348.849 732.315C346.057 732.315 343.294 731.756 340.721 730.673C338.148 729.589 335.817 728.002 333.867 726.004C331.916 724.007 330.385 721.639 329.362 719.041C328.34 716.443 327.848 713.667 327.914 710.876Z"
-      fill="#E067F3"
-    />
-    <path
-      d="M327.914 752.817H369.784C369.851 750.026 369.358 747.25 368.336 744.652C367.314 742.054 365.782 739.687 363.831 737.689C361.881 735.692 359.551 734.105 356.978 733.021C354.405 731.937 351.641 731.379 348.849 731.379C346.057 731.379 343.294 731.937 340.721 733.021C338.148 734.105 335.817 735.692 333.867 737.689C331.916 739.687 330.385 742.054 329.362 744.652C328.34 747.25 327.848 750.026 327.914 752.817Z"
-      fill="#E067F3"
-    />
-    <path
-      d="M375.991 805.439L400.039 786.164L424.087 805.439L424.087 786.164L400.039 766.899L375.991 786.164L351.956 766.899L327.908 786.164L327.908 805.439L351.956 786.164L375.991 805.439Z"
-      fill="#E067F3"
-    />
-    <path
-      d="M472.347 805.439L496.395 786.164L520.443 805.439L520.443 786.164L496.395 766.899L472.347 786.164L448.312 766.899L424.264 786.164L424.264 805.439L448.312 786.164L472.347 805.439Z"
-      fill="#E067F3"
-    />
-    <path
-      d="M434.509 746.112L456.265 746.112C456.265 741.66 458.033 737.391 461.181 734.244C464.328 731.096 468.597 729.328 473.049 729.328C477.5 729.328 481.769 731.096 484.917 734.244C488.065 737.391 489.833 741.66 489.833 746.112L511.588 746.112C511.588 735.891 507.528 726.088 500.3 718.86C493.073 711.633 483.27 707.572 473.049 707.572C462.828 707.572 453.025 711.633 445.797 718.86C438.57 726.088 434.509 735.891 434.509 746.112Z"
-      fill="#E067F3"
-    />
-    <path
-      d="M434.509 707.54L456.265 707.54C456.265 703.088 458.033 698.819 461.181 695.671C464.328 692.524 468.597 690.755 473.049 690.755C477.5 690.755 481.769 692.524 484.917 695.671C488.065 698.819 489.833 703.088 489.833 707.54L511.588 707.54C511.588 697.318 507.528 687.515 500.3 680.288C493.073 673.06 483.27 669 473.049 669C462.828 669 453.025 673.06 445.797 680.288C438.57 687.515 434.509 697.318 434.509 707.54Z"
-      fill="#E067F3"
-    />
-    <path
-      d="M56.1129 175L51.4737 158.273H54.1282L57.3871 171.227H57.5423L60.9318 158.273H63.5618L66.9513 171.235H67.1065L70.3572 158.273H73.0199L68.3725 175H65.8324L62.3121 162.471H62.1815L58.6612 175H56.1129ZM77.5407 175.278C76.7457 175.278 76.0269 175.131 75.3844 174.837C74.7419 174.537 74.2328 174.104 73.8571 173.538C73.4868 172.972 73.3017 172.277 73.3017 171.455C73.3017 170.747 73.4378 170.165 73.71 169.707C73.9823 169.25 74.3498 168.888 74.8127 168.621C75.2755 168.354 75.7928 168.153 76.3645 168.017C76.9363 167.881 77.5189 167.777 78.1124 167.706C78.8638 167.619 79.4737 167.548 79.9419 167.494C80.4102 167.434 80.7505 167.339 80.9629 167.208C81.1752 167.077 81.2814 166.865 81.2814 166.571V166.514C81.2814 165.801 81.08 165.248 80.677 164.856C80.2795 164.464 79.686 164.268 78.8965 164.268C78.0743 164.268 77.4263 164.45 76.9526 164.815C76.4843 165.174 76.1603 165.575 75.9806 166.016L73.6855 165.493C73.9578 164.731 74.3553 164.115 74.878 163.647C75.4062 163.173 76.0133 162.83 76.6994 162.618C77.3855 162.4 78.1069 162.291 78.8638 162.291C79.3648 162.291 79.8957 162.351 80.4565 162.471C81.0228 162.585 81.551 162.798 82.041 163.108C82.5365 163.418 82.9422 163.862 83.258 164.439C83.5738 165.011 83.7317 165.754 83.7317 166.669V175H81.3468V173.285H81.2488C81.0909 173.601 80.854 173.911 80.5382 174.216C80.2224 174.521 79.8167 174.774 79.3212 174.975C78.8257 175.177 78.2322 175.278 77.5407 175.278ZM78.0716 173.317C78.7467 173.317 79.3239 173.184 79.8031 172.917C80.2877 172.65 80.6552 172.302 80.9057 171.872C81.1616 171.436 81.2896 170.971 81.2896 170.475V168.858C81.2025 168.945 81.0337 169.027 80.7832 169.103C80.5382 169.174 80.2578 169.236 79.9419 169.291C79.6261 169.34 79.3185 169.386 79.019 169.43C78.7195 169.468 78.469 169.5 78.2676 169.528C77.7939 169.588 77.361 169.688 76.9689 169.83C76.5823 169.971 76.272 170.176 76.0378 170.442C75.8091 170.704 75.6948 171.052 75.6948 171.488C75.6948 172.092 75.918 172.55 76.3645 172.86C76.811 173.165 77.38 173.317 78.0716 173.317ZM89.4266 158.273V175H86.9845V158.273H89.4266ZM95.1541 158.273V175H92.712V158.273H95.1541ZM103.847 175.253C102.61 175.253 101.546 174.989 100.653 174.461C99.7654 173.927 99.0793 173.179 98.5947 172.215C98.1156 171.246 97.876 170.11 97.876 168.809C97.876 167.524 98.1156 166.391 98.5947 165.411C99.0793 164.431 99.7545 163.666 100.62 163.116C101.492 162.566 102.51 162.291 103.675 162.291C104.383 162.291 105.069 162.408 105.733 162.642C106.398 162.877 106.994 163.244 107.522 163.745C108.05 164.246 108.467 164.897 108.772 165.697C109.076 166.492 109.229 167.459 109.229 168.597V169.462H99.2563V167.633H106.836C106.836 166.99 106.705 166.421 106.444 165.926C106.182 165.425 105.815 165.03 105.341 164.741C104.873 164.453 104.323 164.309 103.691 164.309C103.005 164.309 102.406 164.477 101.894 164.815C101.388 165.147 100.996 165.583 100.718 166.122C100.446 166.655 100.31 167.235 100.31 167.862V169.291C100.31 170.129 100.457 170.843 100.751 171.431C101.05 172.019 101.467 172.468 102.001 172.778C102.534 173.083 103.158 173.236 103.871 173.236C104.334 173.236 104.756 173.17 105.137 173.04C105.518 172.904 105.848 172.702 106.125 172.435C106.403 172.169 106.615 171.839 106.762 171.447L109.074 171.864C108.889 172.544 108.556 173.141 108.077 173.652C107.604 174.159 107.007 174.554 106.289 174.837C105.575 175.114 104.761 175.253 103.847 175.253ZM117.795 162.455V164.415H110.942V162.455H117.795ZM112.78 159.449H115.222V171.316C115.222 171.79 115.293 172.147 115.434 172.386C115.576 172.621 115.758 172.781 115.982 172.868C116.21 172.95 116.458 172.991 116.725 172.991C116.921 172.991 117.092 172.977 117.239 172.95C117.386 172.923 117.501 172.901 117.582 172.885L118.023 174.902C117.882 174.956 117.68 175.011 117.419 175.065C117.158 175.125 116.831 175.158 116.439 175.163C115.796 175.174 115.197 175.06 114.642 174.82C114.087 174.581 113.637 174.21 113.294 173.71C112.951 173.209 112.78 172.58 112.78 171.823V159.449ZM128.192 175H125.513L131.533 158.273H134.449L140.468 175H137.789L133.06 161.311H132.929L128.192 175ZM128.641 168.45H137.332V170.573H128.641V168.45ZM147.182 175.245C146.169 175.245 145.265 174.986 144.47 174.469C143.681 173.946 143.06 173.203 142.608 172.239C142.161 171.27 141.938 170.108 141.938 168.752C141.938 167.396 142.164 166.236 142.616 165.272C143.074 164.309 143.7 163.571 144.495 163.059C145.29 162.547 146.191 162.291 147.198 162.291C147.977 162.291 148.603 162.422 149.077 162.683C149.556 162.939 149.926 163.239 150.188 163.582C150.454 163.925 150.661 164.227 150.808 164.488H150.955V158.273H153.397V175H151.012V173.048H150.808C150.661 173.315 150.449 173.62 150.171 173.963C149.899 174.306 149.523 174.605 149.044 174.861C148.565 175.117 147.944 175.245 147.182 175.245ZM147.721 173.162C148.423 173.162 149.017 172.977 149.501 172.607C149.991 172.231 150.362 171.711 150.612 171.047C150.868 170.383 150.996 169.609 150.996 168.727C150.996 167.856 150.871 167.094 150.62 166.44C150.37 165.787 150.002 165.278 149.518 164.913C149.033 164.548 148.434 164.366 147.721 164.366C146.986 164.366 146.373 164.556 145.883 164.938C145.393 165.319 145.023 165.839 144.772 166.498C144.527 167.156 144.405 167.9 144.405 168.727C144.405 169.566 144.53 170.32 144.781 170.99C145.031 171.659 145.401 172.19 145.891 172.582C146.387 172.969 146.997 173.162 147.721 173.162ZM161.579 175.245C160.566 175.245 159.663 174.986 158.868 174.469C158.078 173.946 157.457 173.203 157.005 172.239C156.559 171.27 156.336 170.108 156.336 168.752C156.336 167.396 156.562 166.236 157.014 165.272C157.471 164.309 158.097 163.571 158.892 163.059C159.687 162.547 160.588 162.291 161.596 162.291C162.374 162.291 163 162.422 163.474 162.683C163.953 162.939 164.324 163.239 164.585 163.582C164.852 163.925 165.059 164.227 165.206 164.488H165.353V158.273H167.795V175H165.41V173.048H165.206C165.059 173.315 164.846 173.62 164.569 173.963C164.296 174.306 163.921 174.605 163.441 174.861C162.962 175.117 162.342 175.245 161.579 175.245ZM162.118 173.162C162.821 173.162 163.414 172.977 163.899 172.607C164.389 172.231 164.759 171.711 165.01 171.047C165.266 170.383 165.394 169.609 165.394 168.727C165.394 167.856 165.268 167.094 165.018 166.44C164.767 165.787 164.4 165.278 163.915 164.913C163.431 164.548 162.832 164.366 162.118 164.366C161.383 164.366 160.771 164.556 160.281 164.938C159.791 165.319 159.42 165.839 159.17 166.498C158.925 167.156 158.802 167.9 158.802 168.727C158.802 169.566 158.928 170.32 159.178 170.99C159.428 171.659 159.799 172.19 160.289 172.582C160.784 172.969 161.394 173.162 162.118 173.162ZM171.28 175V162.455H173.641V164.447H173.771C174 163.772 174.403 163.241 174.98 162.855C175.563 162.463 176.222 162.267 176.957 162.267C177.109 162.267 177.289 162.272 177.496 162.283C177.708 162.294 177.874 162.308 177.994 162.324V164.66C177.896 164.633 177.722 164.603 177.471 164.57C177.221 164.532 176.97 164.513 176.72 164.513C176.143 164.513 175.628 164.635 175.176 164.88C174.73 165.12 174.376 165.455 174.115 165.885C173.853 166.31 173.722 166.794 173.722 167.339V175H171.28ZM185.065 175.253C183.829 175.253 182.765 174.989 181.872 174.461C180.984 173.927 180.298 173.179 179.813 172.215C179.334 171.246 179.095 170.11 179.095 168.809C179.095 167.524 179.334 166.391 179.813 165.411C180.298 164.431 180.973 163.666 181.839 163.116C182.71 162.566 183.728 162.291 184.894 162.291C185.602 162.291 186.288 162.408 186.952 162.642C187.616 162.877 188.213 163.244 188.741 163.745C189.269 164.246 189.685 164.897 189.99 165.697C190.295 166.492 190.448 167.459 190.448 168.597V169.462H180.475V167.633H188.055C188.055 166.99 187.924 166.421 187.663 165.926C187.401 165.425 187.034 165.03 186.56 164.741C186.092 164.453 185.542 164.309 184.91 164.309C184.224 164.309 183.625 164.477 183.113 164.815C182.607 165.147 182.215 165.583 181.937 166.122C181.665 166.655 181.529 167.235 181.529 167.862V169.291C181.529 170.129 181.676 170.843 181.97 171.431C182.269 172.019 182.686 172.468 183.219 172.778C183.753 173.083 184.376 173.236 185.09 173.236C185.553 173.236 185.975 173.17 186.356 173.04C186.737 172.904 187.066 172.702 187.344 172.435C187.622 172.169 187.834 171.839 187.981 171.447L190.293 171.864C190.107 172.544 189.775 173.141 189.296 173.652C188.822 174.159 188.226 174.554 187.507 174.837C186.794 175.114 185.98 175.253 185.065 175.253ZM202.558 165.517L200.345 165.909C200.252 165.626 200.105 165.357 199.904 165.101C199.708 164.845 199.441 164.635 199.103 164.472C198.766 164.309 198.344 164.227 197.837 164.227C197.146 164.227 196.569 164.382 196.106 164.692C195.643 164.997 195.412 165.392 195.412 165.877C195.412 166.296 195.567 166.634 195.877 166.89C196.188 167.145 196.688 167.355 197.38 167.518L199.373 167.976C200.527 168.243 201.388 168.654 201.954 169.209C202.52 169.765 202.803 170.486 202.803 171.374C202.803 172.125 202.585 172.795 202.15 173.383C201.72 173.965 201.118 174.423 200.345 174.755C199.577 175.087 198.687 175.253 197.674 175.253C196.269 175.253 195.123 174.954 194.235 174.355C193.348 173.75 192.803 172.893 192.602 171.782L194.962 171.423C195.109 172.038 195.412 172.503 195.869 172.819C196.326 173.13 196.923 173.285 197.658 173.285C198.458 173.285 199.098 173.119 199.577 172.787C200.056 172.449 200.296 172.038 200.296 171.553C200.296 171.161 200.149 170.832 199.855 170.565C199.566 170.298 199.122 170.097 198.523 169.961L196.4 169.495C195.229 169.228 194.363 168.804 193.803 168.221C193.247 167.638 192.969 166.9 192.969 166.007C192.969 165.267 193.176 164.619 193.59 164.064C194.004 163.508 194.576 163.075 195.305 162.765C196.035 162.449 196.871 162.291 197.813 162.291C199.169 162.291 200.236 162.585 201.015 163.173C201.793 163.756 202.308 164.537 202.558 165.517ZM214.889 165.517L212.676 165.909C212.583 165.626 212.436 165.357 212.235 165.101C212.039 164.845 211.772 164.635 211.434 164.472C211.097 164.309 210.675 164.227 210.168 164.227C209.477 164.227 208.9 164.382 208.437 164.692C207.974 164.997 207.743 165.392 207.743 165.877C207.743 166.296 207.898 166.634 208.208 166.89C208.519 167.145 209.02 167.355 209.711 167.518L211.704 167.976C212.858 168.243 213.719 168.654 214.285 169.209C214.851 169.765 215.134 170.486 215.134 171.374C215.134 172.125 214.917 172.795 214.481 173.383C214.051 173.965 213.449 174.423 212.676 174.755C211.908 175.087 211.018 175.253 210.005 175.253C208.6 175.253 207.454 174.954 206.566 174.355C205.679 173.75 205.134 172.893 204.933 171.782L207.293 171.423C207.44 172.038 207.743 172.503 208.2 172.819C208.657 173.13 209.254 173.285 209.989 173.285C210.789 173.285 211.429 173.119 211.908 172.787C212.387 172.449 212.627 172.038 212.627 171.553C212.627 171.161 212.48 170.832 212.186 170.565C211.897 170.298 211.453 170.097 210.854 169.961L208.731 169.495C207.56 169.228 206.694 168.804 206.134 168.221C205.578 167.638 205.301 166.9 205.301 166.007C205.301 165.267 205.507 164.619 205.921 164.064C206.335 163.508 206.907 163.075 207.636 162.765C208.366 162.449 209.202 162.291 210.144 162.291C211.5 162.291 212.567 162.585 213.346 163.173C214.124 163.756 214.639 164.537 214.889 165.517Z"
-      fill="#BCACE7"
-    />
-    <path
-      d="M62.2244 197.46C63.392 197.469 64.5426 197.682 65.6761 198.099C66.8097 198.517 67.8324 199.199 68.7443 200.145C69.6648 201.091 70.3977 202.369 70.9432 203.98C71.4972 205.582 71.7784 207.577 71.7869 209.963C71.7869 212.256 71.5568 214.297 71.0966 216.087C70.6364 217.868 69.9759 219.372 69.1151 220.599C68.2628 221.827 67.2273 222.76 66.0085 223.399C64.7898 224.038 63.4176 224.358 61.892 224.358C60.3324 224.358 58.9474 224.051 57.7372 223.438C56.527 222.824 55.5426 221.976 54.7841 220.893C54.0256 219.803 53.5526 218.55 53.3651 217.135H57.2642C57.5199 218.26 58.0398 219.172 58.8239 219.871C59.6165 220.561 60.6392 220.906 61.892 220.906C63.8097 220.906 65.3054 220.071 66.3793 218.401C67.4531 216.722 67.9943 214.378 68.0028 211.369H67.7983C67.3551 212.102 66.8011 212.733 66.1364 213.261C65.4801 213.79 64.7429 214.199 63.9247 214.489C63.1065 214.778 62.233 214.923 61.304 214.923C59.7955 214.923 58.4233 214.553 57.1875 213.811C55.9517 213.07 54.9673 212.051 54.2344 210.756C53.5014 209.46 53.1349 207.982 53.1349 206.32C53.1349 204.666 53.5099 203.166 54.2599 201.82C55.0185 200.473 56.0753 199.408 57.4304 198.624C58.794 197.831 60.392 197.443 62.2244 197.46ZM62.2372 200.784C61.2401 200.784 60.3409 201.031 59.5398 201.526C58.7472 202.011 58.1207 202.672 57.6605 203.507C57.2003 204.334 56.9702 205.254 56.9702 206.268C56.9702 207.283 57.1918 208.203 57.6349 209.03C58.0866 209.848 58.7003 210.5 59.4759 210.986C60.2599 211.463 61.1548 211.702 62.1605 211.702C62.9105 211.702 63.6094 211.557 64.2571 211.267C64.9048 210.977 65.4716 210.577 65.9574 210.065C66.4432 209.545 66.8224 208.957 67.0952 208.301C67.3679 207.645 67.5043 206.955 67.5043 206.23C67.5043 205.267 67.2741 204.372 66.8139 203.545C66.3622 202.719 65.7401 202.054 64.9474 201.551C64.1548 201.04 63.2514 200.784 62.2372 200.784ZM79.4734 204.364L83.8072 212.009L88.1793 204.364H92.3597L86.2362 214.182L92.4109 224H88.2305L83.8072 216.662L79.3967 224H75.2035L81.3143 214.182L75.2802 204.364H79.4734ZM105.067 224H96.5913V197.818H105.336C107.901 197.818 110.104 198.342 111.945 199.391C113.786 200.43 115.196 201.926 116.176 203.878C117.165 205.821 117.659 208.152 117.659 210.871C117.659 213.598 117.161 215.942 116.164 217.902C115.175 219.862 113.743 221.371 111.868 222.428C109.993 223.476 107.726 224 105.067 224ZM100.542 220.548H104.85C106.844 220.548 108.502 220.173 109.823 219.423C111.144 218.665 112.132 217.57 112.789 216.138C113.445 214.697 113.773 212.942 113.773 210.871C113.773 208.817 113.445 207.074 112.789 205.642C112.141 204.21 111.174 203.124 109.887 202.382C108.6 201.641 107.002 201.27 105.093 201.27H100.542V220.548ZM122.265 204.364H126.087V225.278C126.087 226.591 125.849 227.699 125.371 228.602C124.903 229.506 124.204 230.192 123.275 230.661C122.354 231.129 121.217 231.364 119.862 231.364C119.725 231.364 119.597 231.364 119.478 231.364C119.35 231.364 119.218 231.359 119.082 231.351V228.065C119.201 228.065 119.308 228.065 119.401 228.065C119.487 228.065 119.585 228.065 119.695 228.065C120.599 228.065 121.251 227.822 121.651 227.337C122.06 226.859 122.265 226.165 122.265 225.253V204.364ZM124.157 201.334C123.492 201.334 122.921 201.112 122.444 200.669C121.975 200.217 121.741 199.68 121.741 199.058C121.741 198.428 121.975 197.891 122.444 197.447C122.921 196.996 123.492 196.77 124.157 196.77C124.822 196.77 125.388 196.996 125.857 197.447C126.335 197.891 126.573 198.428 126.573 199.058C126.573 199.68 126.335 200.217 125.857 200.669C125.388 201.112 124.822 201.334 124.157 201.334ZM135.052 212.341V224H131.23V197.818H135.001V207.56H135.244C135.704 206.503 136.407 205.663 137.353 205.041C138.299 204.419 139.535 204.108 141.061 204.108C142.407 204.108 143.583 204.385 144.589 204.939C145.603 205.493 146.387 206.32 146.941 207.419C147.504 208.51 147.785 209.874 147.785 211.51V224H143.963V211.97C143.963 210.53 143.592 209.413 142.85 208.621C142.109 207.82 141.078 207.419 139.757 207.419C138.853 207.419 138.044 207.611 137.328 207.994C136.62 208.378 136.062 208.94 135.653 209.682C135.252 210.415 135.052 211.301 135.052 212.341ZM161.349 224.396C159.414 224.396 157.748 223.983 156.35 223.156C154.961 222.321 153.887 221.149 153.129 219.641C152.379 218.124 152.004 216.347 152.004 214.31C152.004 212.298 152.379 210.526 153.129 208.991C153.887 207.457 154.944 206.26 156.299 205.399C157.663 204.538 159.257 204.108 161.081 204.108C162.189 204.108 163.262 204.291 164.302 204.658C165.342 205.024 166.275 205.599 167.102 206.384C167.929 207.168 168.581 208.186 169.058 209.439C169.535 210.683 169.774 212.196 169.774 213.977V215.332H154.164V212.469H166.028C166.028 211.463 165.824 210.572 165.414 209.797C165.005 209.013 164.43 208.395 163.689 207.943C162.956 207.491 162.095 207.266 161.106 207.266C160.032 207.266 159.095 207.53 158.294 208.058C157.501 208.578 156.887 209.26 156.453 210.104C156.027 210.939 155.814 211.847 155.814 212.827V215.064C155.814 216.376 156.044 217.493 156.504 218.413C156.973 219.334 157.625 220.037 158.46 220.523C159.295 221 160.271 221.239 161.387 221.239C162.112 221.239 162.772 221.136 163.369 220.932C163.966 220.719 164.481 220.403 164.916 219.986C165.35 219.568 165.683 219.053 165.913 218.439L169.531 219.091C169.241 220.156 168.721 221.089 167.971 221.891C167.23 222.683 166.297 223.301 165.172 223.744C164.055 224.179 162.781 224.396 161.349 224.396ZM183.079 224.358C181.323 224.358 179.755 224.055 178.374 223.45C177.002 222.845 175.915 222.006 175.114 220.932C174.322 219.849 173.896 218.597 173.836 217.173H177.85C177.901 217.949 178.161 218.622 178.63 219.193C179.107 219.756 179.729 220.19 180.496 220.497C181.263 220.804 182.116 220.957 183.053 220.957C184.085 220.957 184.996 220.778 185.789 220.42C186.59 220.062 187.217 219.564 187.668 218.925C188.12 218.277 188.346 217.531 188.346 216.688C188.346 215.81 188.12 215.038 187.668 214.374C187.225 213.7 186.573 213.172 185.712 212.788C184.86 212.405 183.829 212.213 182.619 212.213H180.407V208.991H182.619C183.59 208.991 184.442 208.817 185.175 208.467C185.917 208.118 186.496 207.632 186.914 207.01C187.332 206.379 187.54 205.642 187.54 204.798C187.54 203.989 187.357 203.286 186.991 202.689C186.633 202.084 186.121 201.611 185.457 201.27C184.8 200.929 184.025 200.759 183.13 200.759C182.278 200.759 181.481 200.916 180.739 201.232C180.006 201.538 179.41 201.982 178.95 202.561C178.489 203.132 178.242 203.818 178.208 204.619H174.386C174.428 203.205 174.846 201.96 175.638 200.886C176.44 199.812 177.496 198.973 178.809 198.368C180.121 197.763 181.579 197.46 183.181 197.46C184.86 197.46 186.309 197.788 187.528 198.445C188.755 199.092 189.701 199.957 190.366 201.04C191.039 202.122 191.371 203.307 191.363 204.594C191.371 206.06 190.962 207.304 190.136 208.327C189.317 209.349 188.227 210.036 186.863 210.385V210.589C188.602 210.854 189.948 211.544 190.903 212.661C191.866 213.777 192.343 215.162 192.335 216.815C192.343 218.256 191.942 219.547 191.133 220.689C190.332 221.831 189.237 222.73 187.847 223.386C186.458 224.034 184.869 224.358 183.079 224.358ZM206.001 224H197.525V197.818H206.269C208.835 197.818 211.038 198.342 212.879 199.391C214.719 200.43 216.13 201.926 217.11 203.878C218.099 205.821 218.593 208.152 218.593 210.871C218.593 213.598 218.094 215.942 217.097 217.902C216.109 219.862 214.677 221.371 212.802 222.428C210.927 223.476 208.66 224 206.001 224ZM201.475 220.548H205.783C207.778 220.548 209.435 220.173 210.756 219.423C212.077 218.665 213.066 217.57 213.722 216.138C214.379 214.697 214.707 212.942 214.707 210.871C214.707 208.817 214.379 207.074 213.722 205.642C213.075 204.21 212.107 203.124 210.82 202.382C209.533 201.641 207.935 201.27 206.026 201.27H201.475V220.548ZM227.021 197.818V224H223.199V197.818H227.021ZM235.679 217.339L235.653 212.673H236.318L244.142 204.364H248.719L239.795 213.824H239.195L235.679 217.339ZM232.163 224V197.818H235.986V224H232.163ZM244.564 224L237.533 214.668L240.166 211.996L249.256 224H244.564ZM252.202 204.364H256.025V225.278C256.025 226.591 255.786 227.699 255.309 228.602C254.84 229.506 254.141 230.192 253.212 230.661C252.292 231.129 251.154 231.364 249.799 231.364C249.663 231.364 249.535 231.364 249.415 231.364C249.288 231.364 249.156 231.359 249.019 231.351V228.065C249.138 228.065 249.245 228.065 249.339 228.065C249.424 228.065 249.522 228.065 249.633 228.065C250.536 228.065 251.188 227.822 251.589 227.337C251.998 226.859 252.202 226.165 252.202 225.253V204.364ZM254.094 201.334C253.43 201.334 252.859 201.112 252.381 200.669C251.913 200.217 251.678 199.68 251.678 199.058C251.678 198.428 251.913 197.891 252.381 197.447C252.859 196.996 253.43 196.77 254.094 196.77C254.759 196.77 255.326 196.996 255.795 197.447C256.272 197.891 256.511 198.428 256.511 199.058C256.511 199.68 256.272 200.217 255.795 200.669C255.326 201.112 254.759 201.334 254.094 201.334ZM266.882 224.435C265.637 224.435 264.512 224.205 263.507 223.744C262.501 223.276 261.704 222.598 261.116 221.712C260.537 220.825 260.247 219.739 260.247 218.452C260.247 217.344 260.46 216.432 260.886 215.716C261.312 215 261.887 214.433 262.612 214.016C263.336 213.598 264.146 213.283 265.041 213.07C265.936 212.857 266.848 212.695 267.777 212.584C268.953 212.447 269.907 212.337 270.64 212.251C271.373 212.158 271.906 212.009 272.238 211.804C272.571 211.599 272.737 211.267 272.737 210.807V210.717C272.737 209.601 272.422 208.736 271.791 208.122C271.169 207.509 270.24 207.202 269.004 207.202C267.717 207.202 266.703 207.487 265.961 208.058C265.228 208.621 264.721 209.247 264.44 209.938L260.848 209.119C261.274 207.926 261.896 206.963 262.714 206.23C263.541 205.489 264.491 204.952 265.565 204.619C266.639 204.278 267.768 204.108 268.953 204.108C269.737 204.108 270.568 204.202 271.446 204.389C272.332 204.568 273.159 204.901 273.926 205.386C274.701 205.872 275.336 206.567 275.831 207.47C276.325 208.365 276.572 209.528 276.572 210.96V224H272.839V221.315H272.686C272.439 221.81 272.068 222.295 271.574 222.773C271.079 223.25 270.444 223.646 269.669 223.962C268.893 224.277 267.964 224.435 266.882 224.435ZM267.713 221.366C268.77 221.366 269.673 221.158 270.423 220.74C271.181 220.322 271.757 219.777 272.149 219.104C272.549 218.422 272.75 217.693 272.75 216.918V214.386C272.613 214.523 272.349 214.651 271.957 214.77C271.574 214.881 271.135 214.979 270.64 215.064C270.146 215.141 269.664 215.213 269.196 215.281C268.727 215.341 268.335 215.392 268.02 215.435C267.278 215.528 266.6 215.686 265.987 215.908C265.382 216.129 264.896 216.449 264.529 216.866C264.172 217.276 263.993 217.821 263.993 218.503C263.993 219.449 264.342 220.165 265.041 220.651C265.74 221.128 266.63 221.366 267.713 221.366ZM291.009 204.364V207.432H279.912V204.364H291.009ZM282.955 224V202.088C282.955 200.861 283.223 199.842 283.76 199.033C284.297 198.214 285.009 197.605 285.895 197.205C286.781 196.795 287.744 196.591 288.784 196.591C289.551 196.591 290.207 196.655 290.753 196.783C291.298 196.902 291.703 197.013 291.967 197.115L291.072 200.209C290.893 200.158 290.663 200.098 290.382 200.03C290.101 199.953 289.76 199.915 289.359 199.915C288.43 199.915 287.766 200.145 287.365 200.605C286.973 201.065 286.777 201.73 286.777 202.599V224H282.955ZM307.436 231.364V220.945H307.206C306.976 221.362 306.643 221.839 306.209 222.376C305.783 222.913 305.195 223.382 304.445 223.783C303.695 224.183 302.723 224.384 301.53 224.384C299.945 224.384 298.53 223.979 297.286 223.169C296.05 222.351 295.078 221.187 294.371 219.679C293.672 218.162 293.322 216.342 293.322 214.22C293.322 212.098 293.676 210.283 294.384 208.774C295.099 207.266 296.08 206.111 297.324 205.31C298.568 204.509 299.979 204.108 301.555 204.108C302.774 204.108 303.754 204.312 304.496 204.722C305.246 205.122 305.825 205.591 306.234 206.128C306.652 206.665 306.976 207.138 307.206 207.547H307.526V204.364H311.259V231.364H307.436ZM302.374 221.124C303.473 221.124 304.402 220.834 305.161 220.254C305.928 219.666 306.507 218.852 306.899 217.812C307.3 216.773 307.5 215.562 307.5 214.182C307.5 212.818 307.304 211.625 306.912 210.602C306.52 209.58 305.945 208.783 305.186 208.212C304.428 207.641 303.49 207.355 302.374 207.355C301.223 207.355 300.264 207.653 299.497 208.25C298.73 208.847 298.151 209.661 297.759 210.692C297.375 211.723 297.183 212.886 297.183 214.182C297.183 215.494 297.379 216.675 297.771 217.723C298.163 218.771 298.743 219.602 299.51 220.216C300.286 220.821 301.24 221.124 302.374 221.124ZM325.462 224.358C323.706 224.358 322.138 224.055 320.757 223.45C319.385 222.845 318.298 222.006 317.497 220.932C316.705 219.849 316.278 218.597 316.219 217.173H320.233C320.284 217.949 320.544 218.622 321.013 219.193C321.49 219.756 322.112 220.19 322.879 220.497C323.646 220.804 324.499 220.957 325.436 220.957C326.467 220.957 327.379 220.778 328.172 220.42C328.973 220.062 329.599 219.564 330.051 218.925C330.503 218.277 330.729 217.531 330.729 216.688C330.729 215.81 330.503 215.038 330.051 214.374C329.608 213.7 328.956 213.172 328.095 212.788C327.243 212.405 326.212 212.213 325.001 212.213H322.79V208.991H325.001C325.973 208.991 326.825 208.817 327.558 208.467C328.3 208.118 328.879 207.632 329.297 207.01C329.714 206.379 329.923 205.642 329.923 204.798C329.923 203.989 329.74 203.286 329.374 202.689C329.016 202.084 328.504 201.611 327.839 201.27C327.183 200.929 326.408 200.759 325.513 200.759C324.661 200.759 323.864 200.916 323.122 201.232C322.389 201.538 321.793 201.982 321.332 202.561C320.872 203.132 320.625 203.818 320.591 204.619H316.768C316.811 203.205 317.229 201.96 318.021 200.886C318.822 199.812 319.879 198.973 321.192 198.368C322.504 197.763 323.962 197.46 325.564 197.46C327.243 197.46 328.692 197.788 329.911 198.445C331.138 199.092 332.084 199.957 332.749 201.04C333.422 202.122 333.754 203.307 333.746 204.594C333.754 206.06 333.345 207.304 332.518 208.327C331.7 209.349 330.609 210.036 329.246 210.385V210.589C330.984 210.854 332.331 211.544 333.286 212.661C334.249 213.777 334.726 215.162 334.717 216.815C334.726 218.256 334.325 219.547 333.516 220.689C332.714 221.831 331.619 222.73 330.23 223.386C328.841 224.034 327.251 224.358 325.462 224.358ZM339.601 224V204.364H343.295V207.483H343.5C343.858 206.426 344.489 205.595 345.392 204.99C346.304 204.376 347.335 204.07 348.486 204.07C348.724 204.07 349.006 204.078 349.33 204.095C349.662 204.112 349.922 204.134 350.109 204.159V207.815C349.956 207.773 349.683 207.726 349.291 207.675C348.899 207.615 348.507 207.585 348.115 207.585C347.212 207.585 346.406 207.777 345.699 208.161C345 208.536 344.446 209.06 344.037 209.733C343.628 210.398 343.423 211.156 343.423 212.009V224H339.601ZM353.724 224V197.818H369.96V201.219H357.674V209.196H368.797V212.584H357.674V224H353.724ZM374.511 224V204.364H378.333V224H374.511ZM376.441 201.334C375.777 201.334 375.206 201.112 374.728 200.669C374.26 200.217 374.025 199.68 374.025 199.058C374.025 198.428 374.26 197.891 374.728 197.447C375.206 196.996 375.777 196.77 376.441 196.77C377.106 196.77 377.673 196.996 378.142 197.447C378.619 197.891 378.858 198.428 378.858 199.058C378.858 199.68 378.619 200.217 378.142 200.669C377.673 201.112 377.106 201.334 376.441 201.334ZM391.747 224.396C389.906 224.396 388.3 223.974 386.928 223.131C385.555 222.287 384.49 221.107 383.732 219.589C382.973 218.072 382.594 216.3 382.594 214.271C382.594 212.234 382.973 210.453 383.732 208.928C384.49 207.402 385.555 206.217 386.928 205.374C388.3 204.53 389.906 204.108 391.747 204.108C393.588 204.108 395.195 204.53 396.567 205.374C397.939 206.217 399.004 207.402 399.763 208.928C400.521 210.453 400.901 212.234 400.901 214.271C400.901 216.3 400.521 218.072 399.763 219.589C399.004 221.107 397.939 222.287 396.567 223.131C395.195 223.974 393.588 224.396 391.747 224.396ZM391.76 221.188C392.953 221.188 393.942 220.872 394.726 220.241C395.51 219.611 396.089 218.771 396.464 217.723C396.848 216.675 397.04 215.52 397.04 214.259C397.04 213.006 396.848 211.855 396.464 210.807C396.089 209.75 395.51 208.902 394.726 208.263C393.942 207.624 392.953 207.304 391.76 207.304C390.558 207.304 389.561 207.624 388.768 208.263C387.984 208.902 387.401 209.75 387.017 210.807C386.642 211.855 386.455 213.006 386.455 214.259C386.455 215.52 386.642 216.675 387.017 217.723C387.401 218.771 387.984 219.611 388.768 220.241C389.561 220.872 390.558 221.188 391.76 221.188ZM405.167 204.364H408.99V225.278C408.99 226.591 408.751 227.699 408.274 228.602C407.805 229.506 407.106 230.192 406.177 230.661C405.257 231.129 404.119 231.364 402.764 231.364C402.627 231.364 402.5 231.364 402.38 231.364C402.252 231.364 402.12 231.359 401.984 231.351V228.065C402.103 228.065 402.21 228.065 402.304 228.065C402.389 228.065 402.487 228.065 402.598 228.065C403.501 228.065 404.153 227.822 404.554 227.337C404.963 226.859 405.167 226.165 405.167 225.253V204.364ZM407.059 201.334C406.395 201.334 405.824 201.112 405.346 200.669C404.877 200.217 404.643 199.68 404.643 199.058C404.643 198.428 404.877 197.891 405.346 197.447C405.824 196.996 406.395 196.77 407.059 196.77C407.724 196.77 408.291 196.996 408.76 197.447C409.237 197.891 409.475 198.428 409.475 199.058C409.475 199.68 409.237 200.217 408.76 200.669C408.291 201.112 407.724 201.334 407.059 201.334ZM419.847 224.435C418.602 224.435 417.477 224.205 416.472 223.744C415.466 223.276 414.669 222.598 414.081 221.712C413.501 220.825 413.212 219.739 413.212 218.452C413.212 217.344 413.425 216.432 413.851 215.716C414.277 215 414.852 214.433 415.577 214.016C416.301 213.598 417.111 213.283 418.006 213.07C418.901 212.857 419.813 212.695 420.741 212.584C421.918 212.447 422.872 212.337 423.605 212.251C424.338 212.158 424.871 212.009 425.203 211.804C425.536 211.599 425.702 211.267 425.702 210.807V210.717C425.702 209.601 425.386 208.736 424.756 208.122C424.134 207.509 423.205 207.202 421.969 207.202C420.682 207.202 419.668 207.487 418.926 208.058C418.193 208.621 417.686 209.247 417.405 209.938L413.812 209.119C414.239 207.926 414.861 206.963 415.679 206.23C416.506 205.489 417.456 204.952 418.53 204.619C419.604 204.278 420.733 204.108 421.918 204.108C422.702 204.108 423.533 204.202 424.411 204.389C425.297 204.568 426.124 204.901 426.891 205.386C427.666 205.872 428.301 206.567 428.795 207.47C429.29 208.365 429.537 209.528 429.537 210.96V224H425.804V221.315H425.651C425.403 221.81 425.033 222.295 424.538 222.773C424.044 223.25 423.409 223.646 422.634 223.962C421.858 224.277 420.929 224.435 419.847 224.435ZM420.678 221.366C421.734 221.366 422.638 221.158 423.388 220.74C424.146 220.322 424.722 219.777 425.114 219.104C425.514 218.422 425.714 217.693 425.714 216.918V214.386C425.578 214.523 425.314 214.651 424.922 214.77C424.538 214.881 424.099 214.979 423.605 215.064C423.111 215.141 422.629 215.213 422.161 215.281C421.692 215.341 421.3 215.392 420.984 215.435C420.243 215.528 419.565 215.686 418.952 215.908C418.347 216.129 417.861 216.449 417.494 216.866C417.136 217.276 416.957 217.821 416.957 218.503C416.957 219.449 417.307 220.165 418.006 220.651C418.705 221.128 419.595 221.366 420.678 221.366ZM449.343 209.158L445.878 209.771C445.733 209.328 445.503 208.906 445.188 208.506C444.881 208.105 444.463 207.777 443.935 207.521C443.407 207.266 442.746 207.138 441.953 207.138C440.871 207.138 439.968 207.381 439.243 207.866C438.519 208.344 438.157 208.962 438.157 209.72C438.157 210.376 438.4 210.905 438.885 211.305C439.371 211.706 440.155 212.034 441.238 212.29L444.357 213.006C446.164 213.423 447.51 214.067 448.397 214.936C449.283 215.805 449.726 216.935 449.726 218.324C449.726 219.5 449.385 220.548 448.703 221.469C448.03 222.381 447.088 223.097 445.878 223.616C444.676 224.136 443.283 224.396 441.698 224.396C439.499 224.396 437.705 223.928 436.316 222.99C434.926 222.044 434.074 220.702 433.759 218.963L437.453 218.401C437.684 219.364 438.157 220.092 438.873 220.587C439.588 221.072 440.522 221.315 441.672 221.315C442.925 221.315 443.926 221.055 444.676 220.536C445.426 220.007 445.801 219.364 445.801 218.605C445.801 217.991 445.571 217.476 445.111 217.058C444.659 216.641 443.965 216.325 443.027 216.112L439.703 215.384C437.871 214.966 436.516 214.301 435.638 213.389C434.769 212.477 434.334 211.322 434.334 209.925C434.334 208.766 434.658 207.751 435.306 206.882C435.953 206.013 436.848 205.335 437.99 204.849C439.132 204.355 440.441 204.108 441.915 204.108C444.037 204.108 445.708 204.568 446.926 205.489C448.145 206.401 448.951 207.624 449.343 209.158ZM461.28 224.384C459.695 224.384 458.28 223.979 457.036 223.169C455.8 222.351 454.828 221.187 454.121 219.679C453.422 218.162 453.072 216.342 453.072 214.22C453.072 212.098 453.426 210.283 454.134 208.774C454.849 207.266 455.83 206.111 457.074 205.31C458.318 204.509 459.729 204.108 461.305 204.108C462.524 204.108 463.504 204.312 464.246 204.722C464.996 205.122 465.575 205.591 465.984 206.128C466.402 206.665 466.726 207.138 466.956 207.547H467.186V197.818H471.009V224H467.276V220.945H466.956C466.726 221.362 466.393 221.839 465.959 222.376C465.533 222.913 464.945 223.382 464.195 223.783C463.445 224.183 462.473 224.384 461.28 224.384ZM462.124 221.124C463.223 221.124 464.152 220.834 464.911 220.254C465.678 219.666 466.257 218.852 466.649 217.812C467.05 216.773 467.25 215.562 467.25 214.182C467.25 212.818 467.054 211.625 466.662 210.602C466.27 209.58 465.695 208.783 464.936 208.212C464.178 207.641 463.24 207.355 462.124 207.355C460.973 207.355 460.014 207.653 459.247 208.25C458.48 208.847 457.901 209.661 457.509 210.692C457.125 211.723 456.933 212.886 456.933 214.182C456.933 215.494 457.129 216.675 457.521 217.723C457.913 218.771 458.493 219.602 459.26 220.216C460.036 220.821 460.99 221.124 462.124 221.124ZM485.809 204.364V207.432H474.713V204.364H485.809ZM477.755 224V202.088C477.755 200.861 478.024 199.842 478.561 199.033C479.098 198.214 479.809 197.605 480.696 197.205C481.582 196.795 482.545 196.591 483.585 196.591C484.352 196.591 485.008 196.655 485.554 196.783C486.099 196.902 486.504 197.013 486.768 197.115L485.873 200.209C485.694 200.158 485.464 200.098 485.183 200.03C484.902 199.953 484.561 199.915 484.16 199.915C483.231 199.915 482.566 200.145 482.166 200.605C481.774 201.065 481.578 201.73 481.578 202.599V224H477.755ZM498.817 224.435C496.797 224.426 495.071 223.893 493.64 222.837C492.208 221.78 491.113 220.241 490.354 218.222C489.596 216.202 489.216 213.768 489.216 210.922C489.216 208.084 489.596 205.659 490.354 203.648C491.121 201.636 492.221 200.102 493.652 199.045C495.093 197.989 496.814 197.46 498.817 197.46C500.82 197.46 502.537 197.993 503.969 199.058C505.401 200.115 506.496 201.649 507.255 203.661C508.022 205.663 508.405 208.084 508.405 210.922C508.405 213.777 508.026 216.214 507.267 218.234C506.509 220.246 505.414 221.784 503.982 222.849C502.55 223.906 500.828 224.435 498.817 224.435ZM498.817 221.021C500.59 221.021 501.975 220.156 502.972 218.426C503.978 216.696 504.48 214.195 504.48 210.922C504.48 208.749 504.25 206.912 503.79 205.412C503.338 203.903 502.686 202.761 501.834 201.986C500.99 201.202 499.985 200.81 498.817 200.81C497.053 200.81 495.668 201.679 494.662 203.418C493.657 205.156 493.15 207.658 493.141 210.922C493.141 213.104 493.367 214.949 493.819 216.457C494.279 217.957 494.931 219.095 495.775 219.871C496.618 220.638 497.632 221.021 498.817 221.021Z"
-      fill="white"
-    />
-    <path
-      d="M189.838 551V534.273H200.325V536.445H192.362V541.542H199.778V543.706H192.362V548.827H200.423V551H189.838ZM207.151 551.278C206.356 551.278 205.637 551.131 204.995 550.837C204.352 550.537 203.843 550.104 203.467 549.538C203.097 548.972 202.912 548.277 202.912 547.455C202.912 546.747 203.048 546.165 203.32 545.707C203.593 545.25 203.96 544.888 204.423 544.621C204.886 544.354 205.403 544.153 205.975 544.017C206.547 543.881 207.129 543.777 207.723 543.706C208.474 543.619 209.084 543.548 209.552 543.494C210.021 543.434 210.361 543.339 210.573 543.208C210.786 543.077 210.892 542.865 210.892 542.571V542.514C210.892 541.801 210.69 541.248 210.287 540.856C209.89 540.464 209.296 540.268 208.507 540.268C207.685 540.268 207.037 540.45 206.563 540.815C206.095 541.174 205.771 541.575 205.591 542.016L203.296 541.493C203.568 540.731 203.966 540.115 204.488 539.647C205.017 539.173 205.624 538.83 206.31 538.618C206.996 538.4 207.717 538.291 208.474 538.291C208.975 538.291 209.506 538.351 210.067 538.471C210.633 538.585 211.161 538.798 211.651 539.108C212.147 539.418 212.553 539.862 212.868 540.439C213.184 541.011 213.342 541.754 213.342 542.669V551H210.957V549.285H210.859C210.701 549.601 210.464 549.911 210.149 550.216C209.833 550.521 209.427 550.774 208.932 550.975C208.436 551.177 207.843 551.278 207.151 551.278ZM207.682 549.317C208.357 549.317 208.934 549.184 209.413 548.917C209.898 548.65 210.266 548.302 210.516 547.872C210.772 547.436 210.9 546.971 210.9 546.475V544.858C210.813 544.945 210.644 545.027 210.394 545.103C210.149 545.174 209.868 545.236 209.552 545.291C209.236 545.34 208.929 545.386 208.629 545.43C208.33 545.468 208.079 545.5 207.878 545.528C207.404 545.588 206.971 545.688 206.579 545.83C206.193 545.971 205.882 546.176 205.648 546.442C205.419 546.704 205.305 547.052 205.305 547.488C205.305 548.092 205.528 548.55 205.975 548.86C206.421 549.165 206.99 549.317 207.682 549.317ZM216.595 551V538.455H218.955V540.447H219.086C219.315 539.772 219.718 539.241 220.295 538.855C220.877 538.463 221.536 538.267 222.271 538.267C222.424 538.267 222.603 538.272 222.81 538.283C223.023 538.294 223.189 538.308 223.309 538.324V540.66C223.211 540.633 223.036 540.603 222.786 540.57C222.535 540.532 222.285 540.513 222.034 540.513C221.457 540.513 220.943 540.635 220.491 540.88C220.044 541.12 219.69 541.455 219.429 541.885C219.168 542.31 219.037 542.794 219.037 543.339V551H216.595ZM227.864 543.551V551H225.422V538.455H227.766V540.496H227.921C228.21 539.832 228.662 539.299 229.277 538.896C229.898 538.493 230.679 538.291 231.621 538.291C232.476 538.291 233.225 538.471 233.867 538.83C234.51 539.184 235.008 539.712 235.362 540.415C235.716 541.117 235.893 541.986 235.893 543.02V551H233.451V543.314C233.451 542.405 233.214 541.694 232.74 541.183C232.266 540.665 231.616 540.407 230.788 540.407C230.222 540.407 229.718 540.529 229.277 540.774C228.841 541.019 228.496 541.379 228.24 541.852C227.989 542.321 227.864 542.887 227.864 543.551ZM244.575 551.253C243.339 551.253 242.274 550.989 241.381 550.461C240.494 549.927 239.808 549.179 239.323 548.215C238.844 547.246 238.604 546.11 238.604 544.809C238.604 543.524 238.844 542.391 239.323 541.411C239.808 540.431 240.483 539.666 241.349 539.116C242.22 538.566 243.238 538.291 244.403 538.291C245.111 538.291 245.797 538.408 246.462 538.642C247.126 538.877 247.722 539.244 248.25 539.745C248.779 540.246 249.195 540.897 249.5 541.697C249.805 542.492 249.957 543.459 249.957 544.597V545.462H239.985V543.633H247.564C247.564 542.99 247.434 542.421 247.172 541.926C246.911 541.425 246.543 541.03 246.07 540.741C245.601 540.453 245.051 540.309 244.42 540.309C243.734 540.309 243.135 540.477 242.623 540.815C242.117 541.147 241.725 541.583 241.447 542.122C241.175 542.655 241.038 543.235 241.038 543.862V545.291C241.038 546.129 241.185 546.843 241.479 547.431C241.779 548.019 242.196 548.468 242.729 548.778C243.263 549.083 243.886 549.236 244.6 549.236C245.062 549.236 245.484 549.17 245.866 549.04C246.247 548.904 246.576 548.702 246.854 548.435C247.131 548.169 247.344 547.839 247.491 547.447L249.802 547.864C249.617 548.544 249.285 549.141 248.806 549.652C248.332 550.159 247.736 550.554 247.017 550.837C246.304 551.114 245.49 551.253 244.575 551.253ZM257.363 551.245C256.351 551.245 255.447 550.986 254.652 550.469C253.862 549.946 253.242 549.203 252.79 548.239C252.343 547.27 252.12 546.108 252.12 544.752C252.12 543.396 252.346 542.236 252.798 541.272C253.255 540.309 253.881 539.571 254.676 539.059C255.471 538.547 256.372 538.291 257.38 538.291C258.158 538.291 258.785 538.422 259.258 538.683C259.738 538.939 260.108 539.239 260.369 539.582C260.636 539.925 260.843 540.227 260.99 540.488H261.137V534.273H263.579V551H261.194V549.048H260.99C260.843 549.315 260.631 549.62 260.353 549.963C260.081 550.306 259.705 550.605 259.226 550.861C258.747 551.117 258.126 551.245 257.363 551.245ZM257.903 549.162C258.605 549.162 259.198 548.977 259.683 548.607C260.173 548.231 260.543 547.711 260.794 547.047C261.05 546.383 261.178 545.609 261.178 544.727C261.178 543.856 261.052 543.094 260.802 542.44C260.552 541.787 260.184 541.278 259.699 540.913C259.215 540.548 258.616 540.366 257.903 540.366C257.167 540.366 256.555 540.556 256.065 540.938C255.575 541.319 255.204 541.839 254.954 542.498C254.709 543.156 254.586 543.9 254.586 544.727C254.586 545.566 254.712 546.32 254.962 546.99C255.213 547.659 255.583 548.19 256.073 548.582C256.568 548.969 257.178 549.162 257.903 549.162ZM287.878 542.636C287.878 544.422 287.551 545.958 286.898 547.243C286.244 548.522 285.348 549.508 284.21 550.2C283.078 550.886 281.79 551.229 280.347 551.229C278.899 551.229 277.606 550.886 276.468 550.2C275.335 549.508 274.442 548.52 273.789 547.235C273.135 545.95 272.808 544.417 272.808 542.636C272.808 540.85 273.135 539.318 273.789 538.038C274.442 536.753 275.335 535.767 276.468 535.081C277.606 534.39 278.899 534.044 280.347 534.044C281.79 534.044 283.078 534.39 284.21 535.081C285.348 535.767 286.244 536.753 286.898 538.038C287.551 539.318 287.878 540.85 287.878 542.636ZM285.378 542.636C285.378 541.275 285.158 540.129 284.717 539.198C284.281 538.261 283.682 537.553 282.92 537.074C282.163 536.59 281.305 536.347 280.347 536.347C279.383 536.347 278.523 536.59 277.766 537.074C277.009 537.553 276.41 538.261 275.969 539.198C275.534 540.129 275.316 541.275 275.316 542.636C275.316 543.998 275.534 545.147 275.969 546.083C276.41 547.014 277.009 547.722 277.766 548.207C278.523 548.686 279.383 548.925 280.347 548.925C281.305 548.925 282.163 548.686 282.92 548.207C283.682 547.722 284.281 547.014 284.717 546.083C285.158 545.147 285.378 543.998 285.378 542.636ZM293.27 543.551V551H290.828V538.455H293.172V540.496H293.328C293.616 539.832 294.068 539.299 294.683 538.896C295.304 538.493 296.085 538.291 297.027 538.291C297.882 538.291 298.631 538.471 299.274 538.83C299.916 539.184 300.414 539.712 300.768 540.415C301.122 541.117 301.299 541.986 301.299 543.02V551H298.857V543.314C298.857 542.405 298.62 541.694 298.146 541.183C297.673 540.665 297.022 540.407 296.194 540.407C295.628 540.407 295.124 540.529 294.683 540.774C294.248 541.019 293.902 541.379 293.646 541.852C293.396 542.321 293.27 542.887 293.27 543.551ZM311.966 542.873V544.989H304.55V542.873H311.966ZM329.31 539.712H326.762C326.664 539.168 326.481 538.689 326.214 538.275C325.948 537.861 325.621 537.51 325.234 537.221C324.848 536.933 324.415 536.715 323.936 536.568C323.462 536.421 322.958 536.347 322.425 536.347C321.461 536.347 320.598 536.59 319.835 537.074C319.079 537.559 318.48 538.269 318.039 539.206C317.603 540.143 317.385 541.286 317.385 542.636C317.385 543.998 317.603 545.147 318.039 546.083C318.48 547.02 319.081 547.728 319.844 548.207C320.606 548.686 321.464 548.925 322.416 548.925C322.945 548.925 323.446 548.855 323.919 548.713C324.398 548.566 324.831 548.351 325.218 548.068C325.605 547.785 325.931 547.439 326.198 547.031C326.47 546.617 326.658 546.143 326.762 545.609L329.31 545.618C329.174 546.44 328.91 547.197 328.518 547.888C328.131 548.574 327.633 549.168 327.023 549.669C326.419 550.164 325.727 550.548 324.948 550.82C324.17 551.093 323.32 551.229 322.4 551.229C320.952 551.229 319.661 550.886 318.529 550.2C317.396 549.508 316.503 548.52 315.85 547.235C315.202 545.95 314.878 544.417 314.878 542.636C314.878 540.85 315.204 539.318 315.858 538.038C316.511 536.753 317.404 535.767 318.537 535.081C319.669 534.39 320.957 534.044 322.4 534.044C323.288 534.044 324.115 534.172 324.883 534.428C325.656 534.678 326.351 535.049 326.966 535.539C327.581 536.023 328.09 536.617 328.493 537.319C328.896 538.016 329.168 538.814 329.31 539.712ZM334.576 543.551V551H332.134V534.273H334.543V540.496H334.699C334.993 539.821 335.442 539.285 336.046 538.887C336.651 538.49 337.44 538.291 338.415 538.291C339.275 538.291 340.027 538.468 340.669 538.822C341.317 539.176 341.818 539.704 342.172 540.407C342.531 541.104 342.711 541.975 342.711 543.02V551H340.269V543.314C340.269 542.394 340.032 541.681 339.558 541.174C339.085 540.663 338.426 540.407 337.582 540.407C337.004 540.407 336.487 540.529 336.03 540.774C335.578 541.019 335.221 541.379 334.96 541.852C334.704 542.321 334.576 542.887 334.576 543.551ZM349.621 551.278C348.826 551.278 348.107 551.131 347.464 550.837C346.822 550.537 346.313 550.104 345.937 549.538C345.567 548.972 345.382 548.277 345.382 547.455C345.382 546.747 345.518 546.165 345.79 545.707C346.062 545.25 346.43 544.888 346.893 544.621C347.356 544.354 347.873 544.153 348.445 544.017C349.016 543.881 349.599 543.777 350.192 543.706C350.944 543.619 351.554 543.548 352.022 543.494C352.49 543.434 352.831 543.339 353.043 543.208C353.255 543.077 353.362 542.865 353.362 542.571V542.514C353.362 541.801 353.16 541.248 352.757 540.856C352.36 540.464 351.766 540.268 350.977 540.268C350.154 540.268 349.506 540.45 349.033 540.815C348.564 541.174 348.24 541.575 348.061 542.016L345.766 541.493C346.038 540.731 346.435 540.115 346.958 539.647C347.486 539.173 348.093 538.83 348.779 538.618C349.466 538.4 350.187 538.291 350.944 538.291C351.445 538.291 351.976 538.351 352.537 538.471C353.103 538.585 353.631 538.798 354.121 539.108C354.617 539.418 355.022 539.862 355.338 540.439C355.654 541.011 355.812 541.754 355.812 542.669V551H353.427V549.285H353.329C353.171 549.601 352.934 549.911 352.618 550.216C352.302 550.521 351.897 550.774 351.401 550.975C350.906 551.177 350.312 551.278 349.621 551.278ZM350.152 549.317C350.827 549.317 351.404 549.184 351.883 548.917C352.368 548.65 352.735 548.302 352.986 547.872C353.242 547.436 353.37 546.971 353.37 546.475V544.858C353.283 544.945 353.114 545.027 352.863 545.103C352.618 545.174 352.338 545.236 352.022 545.291C351.706 545.34 351.399 545.386 351.099 545.43C350.8 545.468 350.549 545.5 350.348 545.528C349.874 545.588 349.441 545.688 349.049 545.83C348.662 545.971 348.352 546.176 348.118 546.442C347.889 546.704 347.775 547.052 347.775 547.488C347.775 548.092 347.998 548.55 348.445 548.86C348.891 549.165 349.46 549.317 350.152 549.317ZM359.065 551V538.455H361.507V551H359.065ZM360.298 536.519C359.873 536.519 359.508 536.377 359.203 536.094C358.904 535.806 358.754 535.462 358.754 535.065C358.754 534.662 358.904 534.319 359.203 534.036C359.508 533.747 359.873 533.603 360.298 533.603C360.723 533.603 361.085 533.747 361.384 534.036C361.689 534.319 361.842 534.662 361.842 535.065C361.842 535.462 361.689 535.806 361.384 536.094C361.085 536.377 360.723 536.519 360.298 536.519ZM367.234 543.551V551H364.792V538.455H367.136V540.496H367.291C367.58 539.832 368.032 539.299 368.647 538.896C369.268 538.493 370.049 538.291 370.991 538.291C371.846 538.291 372.595 538.471 373.237 538.83C373.88 539.184 374.378 539.712 374.732 540.415C375.086 541.117 375.263 541.986 375.263 543.02V551H372.821V543.314C372.821 542.405 372.584 541.694 372.11 541.183C371.637 540.665 370.986 540.407 370.158 540.407C369.592 540.407 369.088 540.529 368.647 540.774C368.212 541.019 367.866 541.379 367.61 541.852C367.359 542.321 367.234 542.887 367.234 543.551Z"
-      fill="#BCACE7"
-    />
-    <path
-      d="M162.623 599.636V563.273H164.952V599.636H162.623ZM168.802 575.276C168.688 574.13 168.201 573.24 167.339 572.605C166.477 571.971 165.308 571.653 163.83 571.653C162.826 571.653 161.979 571.795 161.288 572.08C160.596 572.354 160.066 572.738 159.697 573.23C159.337 573.723 159.157 574.281 159.157 574.906C159.138 575.427 159.247 575.882 159.484 576.27C159.73 576.658 160.066 576.994 160.492 577.278C160.918 577.553 161.411 577.795 161.969 578.003C162.528 578.202 163.125 578.372 163.759 578.514L166.373 579.139C167.642 579.423 168.807 579.802 169.867 580.276C170.928 580.749 171.846 581.331 172.623 582.023C173.399 582.714 174.001 583.528 174.427 584.466C174.862 585.403 175.085 586.478 175.094 587.69C175.085 589.471 174.63 591.014 173.731 592.321C172.841 593.618 171.553 594.627 169.867 595.347C168.191 596.057 166.169 596.412 163.802 596.412C161.453 596.412 159.408 596.052 157.665 595.332C155.933 594.613 154.578 593.547 153.603 592.136C152.637 590.716 152.13 588.959 152.083 586.866H158.035C158.101 587.842 158.38 588.656 158.873 589.31C159.375 589.954 160.042 590.441 160.876 590.773C161.719 591.095 162.67 591.256 163.731 591.256C164.772 591.256 165.677 591.104 166.444 590.801C167.22 590.498 167.822 590.077 168.248 589.537C168.674 588.997 168.887 588.377 168.887 587.676C168.887 587.023 168.693 586.473 168.305 586.028C167.926 585.583 167.367 585.205 166.629 584.892C165.899 584.58 165.004 584.295 163.944 584.04L160.776 583.244C158.324 582.648 156.387 581.715 154.967 580.446C153.546 579.177 152.841 577.468 152.85 575.318C152.841 573.557 153.309 572.018 154.256 570.702C155.213 569.385 156.524 568.358 158.191 567.619C159.858 566.881 161.752 566.511 163.873 566.511C166.032 566.511 167.916 566.881 169.526 567.619C171.146 568.358 172.405 569.385 173.305 570.702C174.204 572.018 174.668 573.543 174.697 575.276H168.802ZM189.587 566.511C191.083 566.511 192.522 566.762 193.905 567.264C195.297 567.766 196.537 568.566 197.626 569.665C198.725 570.763 199.591 572.222 200.226 574.04C200.87 575.848 201.196 578.069 201.206 580.702C201.206 583.173 200.927 585.38 200.368 587.321C199.819 589.253 199.028 590.896 197.996 592.25C196.964 593.604 195.718 594.636 194.26 595.347C192.802 596.047 191.173 596.398 189.374 596.398C187.432 596.398 185.718 596.024 184.232 595.276C182.745 594.518 181.547 593.491 180.638 592.193C179.738 590.886 179.194 589.419 179.004 587.79H185.07C185.306 588.85 185.804 589.674 186.561 590.261C187.319 590.839 188.256 591.128 189.374 591.128C191.268 591.128 192.707 590.304 193.692 588.656C194.677 586.999 195.174 584.731 195.183 581.852H194.984C194.549 582.705 193.962 583.438 193.223 584.054C192.484 584.66 191.637 585.129 190.68 585.46C189.733 585.792 188.725 585.957 187.655 585.957C185.941 585.957 184.411 585.555 183.067 584.75C181.722 583.945 180.661 582.837 179.885 581.426C179.108 580.015 178.72 578.405 178.72 576.597C178.711 574.646 179.161 572.908 180.07 571.384C180.979 569.859 182.248 568.666 183.876 567.804C185.515 566.933 187.418 566.502 189.587 566.511ZM189.629 571.341C188.673 571.341 187.821 571.568 187.072 572.023C186.334 572.477 185.747 573.088 185.311 573.855C184.885 574.622 184.677 575.479 184.686 576.426C184.686 577.383 184.894 578.244 185.311 579.011C185.737 579.769 186.315 580.375 187.044 580.83C187.783 581.275 188.625 581.497 189.572 581.497C190.283 581.497 190.941 581.365 191.547 581.099C192.153 580.834 192.679 580.47 193.124 580.006C193.578 579.532 193.933 578.988 194.189 578.372C194.445 577.757 194.572 577.103 194.572 576.412C194.563 575.493 194.345 574.651 193.919 573.884C193.493 573.116 192.906 572.501 192.158 572.037C191.41 571.573 190.567 571.341 189.629 571.341ZM215.993 566.511C217.489 566.511 218.929 566.762 220.311 567.264C221.703 567.766 222.944 568.566 224.033 569.665C225.131 570.763 225.998 572.222 226.632 574.04C227.276 575.848 227.603 578.069 227.612 580.702C227.612 583.173 227.333 585.38 226.774 587.321C226.225 589.253 225.434 590.896 224.402 592.25C223.37 593.604 222.125 594.636 220.666 595.347C219.208 596.047 217.579 596.398 215.78 596.398C213.839 596.398 212.125 596.024 210.638 595.276C209.151 594.518 207.953 593.491 207.044 592.193C206.144 590.886 205.6 589.419 205.411 587.79H211.476C211.713 588.85 212.21 589.674 212.967 590.261C213.725 590.839 214.662 591.128 215.78 591.128C217.674 591.128 219.113 590.304 220.098 588.656C221.083 586.999 221.58 584.731 221.589 581.852H221.391C220.955 582.705 220.368 583.438 219.629 584.054C218.891 584.66 218.043 585.129 217.087 585.46C216.14 585.792 215.131 585.957 214.061 585.957C212.347 585.957 210.818 585.555 209.473 584.75C208.128 583.945 207.068 582.837 206.291 581.426C205.515 580.015 205.126 578.405 205.126 576.597C205.117 574.646 205.567 572.908 206.476 571.384C207.385 569.859 208.654 568.666 210.283 567.804C211.921 566.933 213.824 566.502 215.993 566.511ZM216.036 571.341C215.079 571.341 214.227 571.568 213.479 572.023C212.74 572.477 212.153 573.088 211.717 573.855C211.291 574.622 211.083 575.479 211.092 576.426C211.092 577.383 211.301 578.244 211.717 579.011C212.143 579.769 212.721 580.375 213.45 580.83C214.189 581.275 215.032 581.497 215.979 581.497C216.689 581.497 217.347 581.365 217.953 581.099C218.559 580.834 219.085 580.47 219.53 580.006C219.984 579.532 220.339 578.988 220.595 578.372C220.851 577.757 220.979 577.103 220.979 576.412C220.969 575.493 220.751 574.651 220.325 573.884C219.899 573.116 219.312 572.501 218.564 572.037C217.816 571.573 216.973 571.341 216.036 571.341ZM237.335 592.023L237.179 593.585C237.056 594.835 236.819 596.08 236.469 597.321C236.128 598.571 235.768 599.703 235.389 600.716C235.02 601.729 234.722 602.525 234.494 603.102H230.574C230.716 602.544 230.91 601.762 231.156 600.759C231.402 599.764 231.634 598.647 231.852 597.406C232.07 596.166 232.207 594.902 232.264 593.614L232.335 592.023H237.335ZM252.79 566.511C254.286 566.511 255.725 566.762 257.108 567.264C258.5 567.766 259.741 568.566 260.83 569.665C261.928 570.763 262.795 572.222 263.429 574.04C264.073 575.848 264.4 578.069 264.409 580.702C264.409 583.173 264.13 585.38 263.571 587.321C263.022 589.253 262.231 590.896 261.199 592.25C260.167 593.604 258.921 594.636 257.463 595.347C256.005 596.047 254.376 596.398 252.577 596.398C250.635 596.398 248.921 596.024 247.435 595.276C245.948 594.518 244.75 593.491 243.841 592.193C242.941 590.886 242.397 589.419 242.207 587.79H248.273C248.509 588.85 249.007 589.674 249.764 590.261C250.522 590.839 251.459 591.128 252.577 591.128C254.471 591.128 255.91 590.304 256.895 588.656C257.88 586.999 258.377 584.731 258.386 581.852H258.188C257.752 582.705 257.165 583.438 256.426 584.054C255.688 584.66 254.84 585.129 253.884 585.46C252.937 585.792 251.928 585.957 250.858 585.957C249.144 585.957 247.615 585.555 246.27 584.75C244.925 583.945 243.865 582.837 243.088 581.426C242.312 580.015 241.923 578.405 241.923 576.597C241.914 574.646 242.364 572.908 243.273 571.384C244.182 569.859 245.451 568.666 247.08 567.804C248.718 566.933 250.621 566.502 252.79 566.511ZM252.832 571.341C251.876 571.341 251.024 571.568 250.276 572.023C249.537 572.477 248.95 573.088 248.514 573.855C248.088 574.622 247.88 575.479 247.889 576.426C247.889 577.383 248.098 578.244 248.514 579.011C248.94 579.769 249.518 580.375 250.247 580.83C250.986 581.275 251.829 581.497 252.776 581.497C253.486 581.497 254.144 581.365 254.75 581.099C255.356 580.834 255.882 580.47 256.327 580.006C256.781 579.532 257.136 578.988 257.392 578.372C257.648 577.757 257.776 577.103 257.776 576.412C257.766 575.493 257.548 574.651 257.122 573.884C256.696 573.116 256.109 572.501 255.361 572.037C254.613 571.573 253.77 571.341 252.832 571.341ZM279.196 566.511C280.692 566.511 282.132 566.762 283.514 567.264C284.906 567.766 286.147 568.566 287.236 569.665C288.334 570.763 289.201 572.222 289.835 574.04C290.479 575.848 290.806 578.069 290.815 580.702C290.815 583.173 290.536 585.38 289.977 587.321C289.428 589.253 288.637 590.896 287.605 592.25C286.573 593.604 285.328 594.636 283.869 595.347C282.411 596.047 280.782 596.398 278.983 596.398C277.042 596.398 275.328 596.024 273.841 595.276C272.354 594.518 271.156 593.491 270.247 592.193C269.348 590.886 268.803 589.419 268.614 587.79H274.679C274.916 588.85 275.413 589.674 276.17 590.261C276.928 590.839 277.866 591.128 278.983 591.128C280.877 591.128 282.316 590.304 283.301 588.656C284.286 586.999 284.783 584.731 284.793 581.852H284.594C284.158 582.705 283.571 583.438 282.832 584.054C282.094 584.66 281.246 585.129 280.29 585.46C279.343 585.792 278.334 585.957 277.264 585.957C275.55 585.957 274.021 585.555 272.676 584.75C271.331 583.945 270.271 582.837 269.494 581.426C268.718 580.015 268.33 578.405 268.33 576.597C268.32 574.646 268.77 572.908 269.679 571.384C270.588 569.859 271.857 568.666 273.486 567.804C275.124 566.933 277.027 566.502 279.196 566.511ZM279.239 571.341C278.282 571.341 277.43 571.568 276.682 572.023C275.943 572.477 275.356 573.088 274.92 573.855C274.494 574.622 274.286 575.479 274.295 576.426C274.295 577.383 274.504 578.244 274.92 579.011C275.347 579.769 275.924 580.375 276.653 580.83C277.392 581.275 278.235 581.497 279.182 581.497C279.892 581.497 280.55 581.365 281.156 581.099C281.762 580.834 282.288 580.47 282.733 580.006C283.188 579.532 283.543 578.988 283.798 578.372C284.054 577.757 284.182 577.103 284.182 576.412C284.172 575.493 283.955 574.651 283.528 573.884C283.102 573.116 282.515 572.501 281.767 572.037C281.019 571.573 280.176 571.341 279.239 571.341ZM305.602 566.511C307.098 566.511 308.538 566.762 309.92 567.264C311.313 567.766 312.553 568.566 313.642 569.665C314.741 570.763 315.607 572.222 316.241 574.04C316.885 575.848 317.212 578.069 317.222 580.702C317.222 583.173 316.942 585.38 316.384 587.321C315.834 589.253 315.044 590.896 314.011 592.25C312.979 593.604 311.734 594.636 310.276 595.347C308.817 596.047 307.188 596.398 305.389 596.398C303.448 596.398 301.734 596.024 300.247 595.276C298.76 594.518 297.563 593.491 296.653 592.193C295.754 590.886 295.209 589.419 295.02 587.79H301.085C301.322 588.85 301.819 589.674 302.577 590.261C303.334 590.839 304.272 591.128 305.389 591.128C307.283 591.128 308.723 590.304 309.707 588.656C310.692 586.999 311.189 584.731 311.199 581.852H311C310.564 582.705 309.977 583.438 309.239 584.054C308.5 584.66 307.652 585.129 306.696 585.46C305.749 585.792 304.741 585.957 303.67 585.957C301.956 585.957 300.427 585.555 299.082 584.75C297.738 583.945 296.677 582.837 295.901 581.426C295.124 580.015 294.736 578.405 294.736 576.597C294.726 574.646 295.176 572.908 296.085 571.384C296.994 569.859 298.263 568.666 299.892 567.804C301.53 566.933 303.434 566.502 305.602 566.511ZM305.645 571.341C304.688 571.341 303.836 571.568 303.088 572.023C302.349 572.477 301.762 573.088 301.327 573.855C300.901 574.622 300.692 575.479 300.702 576.426C300.702 577.383 300.91 578.244 301.327 579.011C301.753 579.769 302.33 580.375 303.06 580.83C303.798 581.275 304.641 581.497 305.588 581.497C306.298 581.497 306.956 581.365 307.562 581.099C308.169 580.834 308.694 580.47 309.139 580.006C309.594 579.532 309.949 578.988 310.205 578.372C310.46 577.757 310.588 577.103 310.588 576.412C310.579 575.493 310.361 574.651 309.935 573.884C309.509 573.116 308.921 572.501 308.173 572.037C307.425 571.573 306.582 571.341 305.645 571.341ZM326.945 592.023L326.788 593.585C326.665 594.835 326.429 596.08 326.078 597.321C325.737 598.571 325.377 599.703 324.999 600.716C324.629 601.729 324.331 602.525 324.104 603.102H320.183C320.325 602.544 320.519 601.762 320.766 600.759C321.012 599.764 321.244 598.647 321.462 597.406C321.679 596.166 321.817 594.902 321.874 593.614L321.945 592.023H326.945ZM342.655 596.639C340.212 596.63 338.109 596.028 336.348 594.835C334.596 593.642 333.247 591.914 332.3 589.651C331.362 587.387 330.898 584.665 330.908 581.483C330.908 578.311 331.376 575.607 332.314 573.372C333.261 571.137 334.61 569.437 336.362 568.273C338.124 567.098 340.221 566.511 342.655 566.511C345.089 566.511 347.181 567.098 348.933 568.273C350.695 569.447 352.049 571.152 352.996 573.386C353.943 575.612 354.411 578.311 354.402 581.483C354.402 584.674 353.929 587.402 352.982 589.665C352.044 591.928 350.699 593.656 348.947 594.849C347.196 596.043 345.098 596.639 342.655 596.639ZM342.655 591.54C344.321 591.54 345.652 590.702 346.646 589.026C347.641 587.349 348.133 584.835 348.124 581.483C348.124 579.277 347.896 577.439 347.442 575.972C346.997 574.504 346.362 573.401 345.538 572.662C344.724 571.923 343.763 571.554 342.655 571.554C340.998 571.554 339.672 572.383 338.678 574.04C337.683 575.697 337.181 578.178 337.172 581.483C337.172 583.718 337.394 585.583 337.839 587.08C338.294 588.566 338.933 589.684 339.757 590.432C340.581 591.17 341.547 591.54 342.655 591.54ZM370.194 596.639C367.751 596.63 365.648 596.028 363.887 594.835C362.135 593.642 360.786 591.914 359.839 589.651C358.901 587.387 358.437 584.665 358.447 581.483C358.447 578.311 358.915 575.607 359.853 573.372C360.8 571.137 362.149 569.437 363.901 568.273C365.663 567.098 367.76 566.511 370.194 566.511C372.628 566.511 374.72 567.098 376.472 568.273C378.234 569.447 379.588 571.152 380.535 573.386C381.482 575.612 381.951 578.311 381.941 581.483C381.941 584.674 381.468 587.402 380.521 589.665C379.583 591.928 378.238 593.656 376.487 594.849C374.735 596.043 372.637 596.639 370.194 596.639ZM370.194 591.54C371.861 591.54 373.191 590.702 374.185 589.026C375.18 587.349 375.672 584.835 375.663 581.483C375.663 579.277 375.435 577.439 374.981 575.972C374.536 574.504 373.901 573.401 373.077 572.662C372.263 571.923 371.302 571.554 370.194 571.554C368.537 571.554 367.211 572.383 366.217 574.04C365.222 575.697 364.72 578.178 364.711 581.483C364.711 583.718 364.933 585.583 365.379 587.08C365.833 588.566 366.472 589.684 367.296 590.432C368.12 591.17 369.086 591.54 370.194 591.54ZM397.733 596.639C395.29 596.63 393.188 596.028 391.426 594.835C389.674 593.642 388.325 591.914 387.378 589.651C386.44 587.387 385.976 584.665 385.986 581.483C385.986 578.311 386.455 575.607 387.392 573.372C388.339 571.137 389.688 569.437 391.44 568.273C393.202 567.098 395.299 566.511 397.733 566.511C400.167 566.511 402.259 567.098 404.011 568.273C405.773 569.447 407.127 571.152 408.074 573.386C409.021 575.612 409.49 578.311 409.48 581.483C409.48 584.674 409.007 587.402 408.06 589.665C407.122 591.928 405.777 593.656 404.026 594.849C402.274 596.043 400.176 596.639 397.733 596.639ZM397.733 591.54C399.4 591.54 400.73 590.702 401.724 589.026C402.719 587.349 403.211 584.835 403.202 581.483C403.202 579.277 402.974 577.439 402.52 575.972C402.075 574.504 401.44 573.401 400.616 572.662C399.802 571.923 398.841 571.554 397.733 571.554C396.076 571.554 394.75 572.383 393.756 574.04C392.761 575.697 392.259 578.178 392.25 581.483C392.25 583.718 392.473 585.583 392.918 587.08C393.372 588.566 394.011 589.684 394.835 590.432C395.659 591.17 396.625 591.54 397.733 591.54Z"
-      fill="white"
-    />
-    <circle
-      cx="495"
-      cy="78"
-      r="26.5"
-      stroke="black"
-      stroke-opacity="0.12"
-      stroke-width="17"
-    />
-    <rect x="144" width="274" height="30" fill="#3B0FBB" />
-    <path d="M520 271L520 358" stroke="white" stroke-opacity="0.45" />
-    <rect
-      x="534.115"
-      y="532.897"
-      width="63.1026"
-      height="27.4359"
-      rx="9"
-      transform="rotate(90 534.115 532.897)"
-      fill="#380DB4"
-    />
-    <path
-      d="M515.596 553.404L516.986 553.404L516.986 547.436L519.738 547.436L519.738 553.195L521.114 553.195L521.114 547.436L523.809 547.436L523.809 553.404L525.199 553.404L525.199 545.93L515.596 545.93L515.596 553.404Z"
-      fill="white"
-    />
-    <path
-      d="M515.596 555.553L518.09 556.627L518.09 561.564L515.596 562.638L515.596 564.339L525.199 560.169L525.199 558.05L515.596 553.88L515.596 555.553ZM522.404 558.496L523.766 559.04L523.766 559.151L522.404 559.695L519.48 560.964L519.48 557.227L522.404 558.496Z"
-      fill="white"
-    />
-    <path
-      d="M515.596 566.542L519.323 566.542L519.323 568.257C519.323 569.429 519.007 569.861 518.248 570.28L515.596 571.772L515.596 573.501L518.377 571.939C518.979 571.604 519.409 571.27 519.609 570.544L519.71 570.544C519.81 572.302 520.784 573.32 522.375 573.32C524.167 573.32 525.199 572.009 525.199 569.889L525.199 565.036L515.596 565.036L515.596 566.542ZM523.809 566.542L523.809 569.875C523.809 571.13 523.336 571.758 522.16 571.758C520.956 571.758 520.512 571.13 520.512 569.875L520.512 566.542L523.809 566.542Z"
-      fill="white"
-    />
-    <path
-      d="M515.596 575.912L521.229 575.912L522.891 575.856L522.891 575.967L521.2 577.139L515.596 581.35L515.596 582.968L525.199 582.968L525.199 581.462L519.523 581.462L517.789 581.532L517.789 581.42L519.581 580.207L525.199 576.023L525.199 574.406L515.596 574.406L515.596 575.912Z"
-      fill="white"
-    />
-    <path
-      d="M511.58 388.904C511.58 387.919 511.681 387.005 511.881 386.166C512.084 385.328 512.37 384.597 512.748 383.975C513.124 383.353 513.592 382.866 514.15 382.52C514.71 382.174 515.354 382 516.087 382C516.162 382 516.233 382 516.308 382C516.382 382 516.445 382.008 516.502 382.025L516.502 385.789C516.466 385.789 516.424 385.789 516.377 385.789C516.332 385.789 516.29 385.789 516.251 385.789C515.81 385.806 515.452 385.961 515.178 386.256C514.904 386.549 514.707 386.94 514.588 387.424C514.469 387.908 514.409 388.42 514.409 388.954C514.409 389.455 514.445 389.944 514.519 390.42C514.594 390.895 514.74 391.295 514.96 391.613C515.181 391.933 515.473 392.094 515.84 392.094C516.299 392.094 516.636 391.902 516.856 391.523C517.077 391.143 517.25 390.62 517.378 389.953C517.506 389.286 517.652 388.538 517.819 387.708C517.948 387.016 518.112 386.355 518.314 385.722C518.517 385.092 518.776 384.523 519.098 384.023C519.42 383.522 519.832 383.122 520.336 382.83C520.839 382.535 521.468 382.388 522.22 382.388C523.007 382.388 523.686 382.551 524.256 382.881C524.825 383.21 525.29 383.677 525.657 384.281C526.024 384.886 526.295 385.601 526.468 386.422C526.641 387.244 526.73 388.13 526.73 389.081C526.73 389.964 526.644 390.794 526.468 391.573C526.295 392.352 526.036 393.028 525.699 393.596C525.359 394.167 524.933 394.62 524.42 394.957C523.907 395.295 523.329 395.464 522.688 395.464C522.577 395.464 522.476 395.464 522.387 395.464C522.294 395.464 522.223 395.455 522.166 395.439L522.166 391.677L522.36 391.677C522.709 391.677 522.992 391.559 523.212 391.326C523.433 391.092 523.603 390.771 523.722 390.366C523.841 389.958 523.901 389.497 523.901 388.977C523.901 388.718 523.883 388.422 523.847 388.093C523.812 387.764 523.746 387.449 523.654 387.145C523.561 386.841 523.433 386.6 523.269 386.419C523.105 386.239 522.893 386.147 522.637 386.147C522.306 386.147 522.047 386.31 521.853 386.639C521.659 386.968 521.51 387.404 521.4 387.95C521.29 388.496 521.17 389.086 521.042 389.728C520.878 390.454 520.708 391.182 520.532 391.908C520.359 392.634 520.124 393.3 519.832 393.905C519.54 394.51 519.116 394.986 518.568 395.332C518.019 395.678 517.283 395.849 516.368 395.849C515.488 395.849 514.74 395.68 514.126 395.343C513.512 395.005 513.017 394.527 512.641 393.903C512.265 393.281 511.994 392.544 511.83 391.697C511.663 390.859 511.58 389.925 511.58 388.904Z"
-      fill="white"
-    />
-    <path
-      d="M511.58 403.616C511.58 401.99 512.021 400.733 512.9 399.841C513.78 398.949 515.163 398.505 517.053 398.505L526.375 398.505L526.375 402.294L517.715 402.294C517.181 402.294 516.737 402.353 516.382 402.477C516.025 402.598 515.736 402.781 515.515 403.023C515.294 403.265 515.133 403.554 515.035 403.892C514.937 404.23 514.883 404.615 514.883 405.045C514.883 405.667 515.029 406.235 515.324 406.745C515.616 407.254 516.025 407.665 516.546 407.977C517.068 408.289 517.661 408.444 518.32 408.444L526.378 408.444L526.378 412.258L511.914 412.258L511.914 409.144L514.114 408.86L514.114 408.677C513.544 408.261 513.073 407.794 512.698 407.276C512.322 406.756 512.042 406.182 511.86 405.552C511.672 404.919 511.58 404.275 511.58 403.616Z"
-      fill="white"
-    />
-    <path
-      d="M506.879 415.682L526.376 415.682L526.376 418.771L524.286 419.109L524.286 419.291C525.165 419.829 525.788 420.507 526.155 421.328C526.522 422.15 526.704 423.044 526.704 424.015C526.704 425.278 526.438 426.372 525.905 427.298C525.374 428.223 524.548 428.938 523.43 429.438C522.312 429.939 520.881 430.19 519.14 430.19C517.435 430.19 516.022 429.934 514.904 429.424C513.786 428.915 512.951 428.209 512.403 427.309C511.854 426.409 511.577 425.354 511.577 424.144C511.577 423.539 511.652 422.954 511.798 422.391C511.944 421.829 512.17 421.314 512.471 420.847C512.772 420.38 513.145 419.981 513.586 419.654L513.586 419.5L506.876 419.5L506.876 415.682L506.879 415.682ZM514.853 422.923C514.853 423.702 515.005 424.341 515.306 424.845C515.608 425.345 516.067 425.717 516.681 425.961C517.295 426.203 518.022 426.324 518.866 426.324L519.415 426.324C520.294 426.324 521.036 426.203 521.642 425.961C522.247 425.719 522.7 425.348 523.004 424.845C523.305 424.341 523.457 423.694 523.457 422.898C523.457 422.155 523.293 421.525 522.962 421.016C522.631 420.507 522.16 420.116 521.546 419.848C520.932 419.578 520.202 419.446 519.361 419.446L518.976 419.446C518.353 419.446 517.781 419.525 517.259 419.68C516.737 419.834 516.296 420.057 515.939 420.341C515.581 420.625 515.309 420.979 515.128 421.404C514.943 421.829 514.853 422.335 514.853 422.923Z"
-      fill="white"
-    />
-    <path
-      d="M511.58 440.126C511.58 438.483 511.851 437.085 512.391 435.934C512.93 434.783 513.762 433.906 514.88 433.301C515.998 432.696 517.42 432.392 519.143 432.392C520.866 432.392 522.288 432.693 523.406 433.301C524.524 433.906 525.353 434.775 525.896 435.909C526.435 437.042 526.707 438.412 526.707 440.021C526.707 441.597 526.441 442.933 525.908 444.03C525.377 445.127 524.566 445.96 523.475 446.52C522.384 447.083 521.015 447.364 519.364 447.364L518.237 447.364L518.237 436.311C517.393 436.345 516.693 436.511 516.132 436.803C515.572 437.096 515.163 437.526 514.907 438.089C514.651 438.651 514.522 439.346 514.522 440.179C514.522 440.593 514.573 440.992 514.674 441.372C514.776 441.752 514.931 442.098 515.142 442.41C515.354 442.722 515.613 442.981 515.927 443.189C516.237 443.397 516.594 443.51 517 443.527L517 447.341C516.156 447.341 515.396 447.164 514.716 446.81C514.036 446.455 513.47 445.966 513.011 445.344C512.552 444.722 512.2 443.968 511.953 443.085C511.705 442.202 511.58 441.217 511.58 440.126ZM520.738 436.364L520.738 443.369C521.251 443.369 521.701 443.288 522.086 443.122C522.47 442.956 522.792 442.728 523.049 442.435C523.305 442.14 523.493 441.794 523.612 441.397C523.731 441 523.791 440.559 523.791 440.075C523.791 439.332 523.678 438.696 523.448 438.168C523.219 437.639 522.876 437.228 522.417 436.935C521.955 436.64 521.397 436.449 520.738 436.364Z"
-      fill="white"
-    />
-    <path
-      d="M511.911 450.169L526.376 450.169L526.376 453.283L524.012 453.595L524.012 453.778C524.506 453.95 524.966 454.203 525.386 454.529C525.806 454.858 526.137 455.261 526.376 455.736C526.614 456.211 526.733 456.751 526.733 457.359C526.733 457.671 526.707 457.956 526.65 458.214C526.596 458.473 526.531 458.698 526.456 458.89L522.935 458.89L522.935 457.384C522.935 456.797 522.84 456.29 522.646 455.865C522.452 455.441 522.178 455.086 521.82 454.802C521.463 454.518 521.045 454.31 520.568 454.18C520.091 454.051 519.552 453.986 518.947 453.986L511.908 453.986L511.908 450.169L511.911 450.169Z"
-      fill="white"
-    />
-    <path
-      d="M511.58 465.866C511.58 464.915 511.735 464.136 512.048 463.531C512.358 462.926 512.823 462.479 513.437 462.195C514.051 461.911 514.817 461.767 515.733 461.767L523.35 461.767L523.35 459.666L526.376 459.666L526.376 461.922L530.582 462.648L530.582 465.579L526.376 465.579L526.376 468.64L523.35 468.64L523.35 465.579L516.311 465.579C515.762 465.579 515.345 465.686 515.059 465.903C514.776 466.119 514.633 466.53 514.633 467.135L514.633 468.64L512.021 468.64C511.947 468.415 511.878 468.139 511.815 467.81C511.753 467.481 511.696 467.143 511.651 466.797C511.604 466.454 511.58 466.145 511.58 465.866Z"
-      fill="white"
-    />
-    <path
-      d="M511.58 478.323C511.58 476.68 511.851 475.282 512.391 474.131C512.93 472.981 513.762 472.103 514.88 471.498C515.998 470.893 517.42 470.589 519.143 470.589C520.866 470.589 522.288 470.893 523.406 471.498C524.524 472.103 525.353 472.972 525.896 474.106C526.435 475.24 526.707 476.61 526.707 478.219C526.707 479.791 526.441 481.13 525.908 482.228C525.377 483.325 524.566 484.157 523.475 484.72C522.384 485.283 521.015 485.564 519.364 485.564L518.237 485.564L518.237 474.511C517.393 474.545 516.693 474.711 516.132 475.003C515.572 475.296 515.163 475.726 514.907 476.289C514.651 476.852 514.522 477.546 514.522 478.379C514.522 478.795 514.573 479.192 514.674 479.572C514.776 479.952 514.931 480.298 515.142 480.61C515.354 480.922 515.613 481.181 515.927 481.389C516.237 481.597 516.594 481.71 517 481.727L517 485.541C516.156 485.541 515.396 485.364 514.716 485.01C514.036 484.655 513.47 484.166 513.011 483.544C512.552 482.922 512.2 482.168 511.953 481.288C511.705 480.399 511.58 479.412 511.58 478.323ZM520.738 474.559L520.738 481.566C521.251 481.566 521.701 481.485 522.086 481.319C522.47 481.156 522.792 480.925 523.049 480.632C523.305 480.337 523.493 479.991 523.612 479.594C523.731 479.195 523.791 478.756 523.791 478.272C523.791 477.53 523.678 476.894 523.448 476.365C523.219 475.836 522.876 475.425 522.417 475.133C521.955 474.837 521.397 474.646 520.738 474.559Z"
-      fill="white"
-    />
-    <path
-      d="M511.58 492.673C511.58 492.068 511.645 491.463 511.774 490.855C511.902 490.251 512.113 489.71 512.406 489.235C512.698 488.76 513.115 488.374 513.658 488.082C514.197 487.786 514.889 487.64 515.733 487.64C516.815 487.64 517.688 487.882 518.359 488.366C519.027 488.85 519.531 489.55 519.87 490.467C520.21 491.384 520.434 492.47 520.544 493.725C520.655 494.98 520.708 496.358 520.708 497.863L521.561 497.863C522.056 497.863 522.464 497.765 522.783 497.565C523.105 497.365 523.356 497.056 523.54 496.631C523.722 496.206 523.815 495.641 523.815 494.932C523.815 494.31 523.746 493.778 523.609 493.337C523.472 492.895 523.293 492.557 523.072 492.324C522.852 492.09 522.595 491.972 522.303 491.972L521.972 491.972L521.972 488.236C522.083 488.219 522.175 488.205 522.247 488.197C522.321 488.189 522.41 488.183 522.521 488.183C523.364 488.183 524.101 488.45 524.733 488.987C525.365 489.525 525.851 490.307 526.191 491.336C526.531 492.366 526.7 493.598 526.7 495.033C526.7 496.468 526.536 497.675 526.206 498.654C525.875 499.63 525.362 500.378 524.664 500.898C523.967 501.416 523.078 501.678 521.996 501.678L515.643 501.678C515.312 501.678 515.062 501.759 514.886 501.925C514.713 502.088 514.624 502.294 514.624 502.536L514.624 503.624L512.039 503.624C511.946 503.433 511.851 503.132 511.75 502.716C511.648 502.299 511.598 501.816 511.598 501.261C511.598 500.62 511.684 500.094 511.86 499.678C512.033 499.261 512.271 498.938 512.576 498.704C512.88 498.471 513.231 498.302 513.634 498.198L513.634 497.99C513.231 497.627 512.868 497.179 512.549 496.653C512.227 496.124 511.985 495.528 511.821 494.861C511.663 494.197 511.58 493.466 511.58 492.673ZM514.549 493.812C514.549 494.313 514.618 494.802 514.755 495.278C514.892 495.753 515.086 496.181 515.333 496.563C515.581 496.943 515.894 497.255 516.269 497.497C516.645 497.739 517.053 497.86 517.492 497.86L518.287 497.86C518.287 496.51 518.219 495.359 518.082 494.408C517.945 493.458 517.715 492.726 517.393 492.217C517.071 491.708 516.627 491.452 516.061 491.452C515.694 491.452 515.402 491.556 515.181 491.764C514.96 491.972 514.799 492.256 514.701 492.619C514.6 492.982 514.549 493.382 514.549 493.812Z"
-      fill="white"
-    />
-    <path
-      d="M511.911 505.411L526.376 505.411L526.376 508.525L524.202 508.837L524.202 509.02C524.772 509.383 525.243 509.819 525.618 510.331C525.994 510.84 526.268 511.392 526.444 511.979C526.617 512.567 526.707 513.172 526.707 513.797C526.707 514.852 526.51 515.743 526.116 516.469C525.723 517.195 525.085 517.749 524.205 518.129L524.205 518.312C524.775 518.692 525.246 519.139 525.621 519.648C525.997 520.157 526.271 520.72 526.447 521.336C526.62 521.949 526.709 522.577 526.709 523.218C526.709 524.239 526.525 525.109 526.161 525.826C525.794 526.543 525.207 527.1 524.402 527.5C523.594 527.896 522.569 528.096 521.322 528.096L511.917 528.096L511.917 524.256L520.634 524.256C521.147 524.256 521.588 524.2 521.954 524.087C522.321 523.975 522.613 523.814 522.834 523.606C523.055 523.398 523.21 523.134 523.302 522.816C523.394 522.495 523.439 522.155 523.439 521.792C523.439 521.221 523.296 520.697 523.013 520.222C522.73 519.747 522.324 519.367 521.802 519.08C521.281 518.796 520.687 518.652 520.029 518.652L511.917 518.652L511.917 514.863L520.634 514.863C521.147 514.863 521.588 514.801 521.954 514.68C522.321 514.559 522.613 514.39 522.834 514.174C523.055 513.957 523.21 513.698 523.302 513.394C523.394 513.091 523.439 512.742 523.439 512.342C523.439 511.771 523.296 511.248 523.013 510.773C522.73 510.297 522.324 509.92 521.802 509.645C521.281 509.369 520.687 509.228 520.029 509.228L511.917 509.228L511.917 505.411L511.911 505.411Z"
-      fill="white"
-    />
-    <path d="M-7 259L37.5 303.5" stroke="white" stroke-width="5" />
-    <path d="M-7 303.5L37.5 348" stroke="white" stroke-width="5" />
-    <path d="M-7 348L37.5 392.5" stroke="white" stroke-width="5" />
-    <path d="M-7 392.5L37.5 437" stroke="white" stroke-width="5" />
-    <path d="M-7 437L37.5 481.5" stroke="white" stroke-width="5" />
-    <path d="M-7 481.5L37.5 526" stroke="white" stroke-width="5" />
-    <path d="M-7 526L37.5 570.5" stroke="white" stroke-width="5" />
-  </g>
-  <defs>
-    <clipPath id="clip0_41_9577">
-      <rect width="565" height="783" fill="white" />
-    </clipPath>
-  </defs>
-</svg>;
-`;
-};
